@@ -20,7 +20,6 @@ import (
 	"bytes"
 	"compress/gzip"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -48,8 +47,8 @@ func TestBuildWithLocalTar(t *testing.T) {
 	nonExistingDockerfile := "Dockerfile_non_existing"
 
 	files := map[string]string{
-		validDockerfile:   "FROM debian:9.11\nRUN echo \"valid\"",
-		invalidDockerfile: "FROM debian:9.11\nRUN echo \"invalid\"",
+		validDockerfile:   "FROM debian:10.13\nRUN echo \"valid\"",
+		invalidDockerfile: "FROM debian:10.13\nRUN echo \"invalid\"",
 	}
 
 	if err := testutil.SetupFiles(testDir, files); err != nil {
@@ -145,7 +144,7 @@ func TestBuildWithLocalTar(t *testing.T) {
 }
 
 func getSHAFromFilePath(f string) (string, error) {
-	data, err := ioutil.ReadFile(f)
+	data, err := os.ReadFile(f)
 	if err != nil {
 		return "", err
 	}
