@@ -13,13 +13,14 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package commands
 
 import (
 	"archive/tar"
 	"fmt"
 	"io"
-	"io/ioutil"
+	"io/fs"
 	"os"
 	"path/filepath"
 	"strings"
@@ -50,18 +51,35 @@ var copyTests = []struct {
 		sourcesAndDest: []string{"foo", "tempCopyExecuteTest"},
 		expectedDest:   []string{"tempCopyExecuteTest"},
 	},
+	{
+		name:           "copy f* into tempCopyExecuteTest",
+		sourcesAndDest: []string{"foo*", "tempCopyExecuteTest"},
+		expectedDest:   []string{"tempCopyExecuteTest"},
+	},
+	{
+		name:           "copy fo? into tempCopyExecuteTest",
+		sourcesAndDest: []string{"fo?", "tempCopyExecuteTest"},
+		expectedDest:   []string{"tempCopyExecuteTest"},
+	},
+	{
+		name:           "copy f[o][osp] into tempCopyExecuteTest",
+		sourcesAndDest: []string{"f[o][osp]", "tempCopyExecuteTest"},
+		expectedDest:   []string{"tempCopyExecuteTest"},
+	},
+	{
+		name:           "Copy into several to-be-created directories",
+		sourcesAndDest: []string{"f[o][osp]", "tempCopyExecuteTest/foo/bar"},
+		expectedDest:   []string{"bar"},
+	},
 }
 
-func setupTestTemp() string {
-	tempDir, err := ioutil.TempDir("", "")
-	if err != nil {
-		logrus.Fatalf("error creating temp dir %s", err)
-	}
+func setupTestTemp(t *testing.T) string {
+	tempDir := t.TempDir()
 	logrus.Debugf("Tempdir: %s", tempDir)
 
 	srcPath, err := filepath.Abs("../../integration/context")
 	if err != nil {
-		logrus.Fatalf("error getting abs path %s", srcPath)
+		logrus.Fatalf("Error getting abs path %s", srcPath)
 	}
 	cperr := filepath.Walk(srcPath,
 		func(path string, info os.FileInfo, err error) error {
@@ -101,16 +119,34 @@ func setupTestTemp() string {
 			return nil
 		})
 	if cperr != nil {
-		logrus.Fatalf("error populating temp dir %s", cperr)
+		logrus.Fatalf("Error populating temp dir %s", cperr)
 	}
 
 	return tempDir
 }
 
-func Test_CachingCopyCommand_ExecuteCommand(t *testing.T) {
-	tempDir := setupTestTemp()
+func readDirectory(dirName string) ([]fs.FileInfo, error) {
+	entries, err := os.ReadDir(dirName)
+	if err != nil {
+		return nil, err
+	}
 
-	tarContent, err := prepareTarFixture([]string{"foo.txt"})
+	testDir := make([]fs.FileInfo, 0, len(entries))
+
+	for _, entry := range entries {
+		info, err := entry.Info()
+		if err != nil {
+			return nil, err
+		}
+		testDir = append(testDir, info)
+	}
+	return testDir, err
+}
+
+func Test_CachingCopyCommand_ExecuteCommand(t *testing.T) {
+	tempDir := setupTestTemp(t)
+
+	tarContent, err := prepareTarFixture(t, []string{"foo.txt"})
 	if err != nil {
 		t.Errorf("couldn't prepare tar fixture %v", err)
 	}
@@ -119,7 +155,7 @@ func Test_CachingCopyCommand_ExecuteCommand(t *testing.T) {
 	buildArgs := &dockerfile.BuildArgs{}
 
 	type testCase struct {
-		desctiption    string
+		description    string
 		expectLayer    bool
 		expectErr      bool
 		count          *int
@@ -130,7 +166,7 @@ func Test_CachingCopyCommand_ExecuteCommand(t *testing.T) {
 	}
 	testCases := []testCase{
 		func() testCase {
-			err = ioutil.WriteFile(filepath.Join(tempDir, "foo.txt"), []byte("meow"), 0644)
+			err = os.WriteFile(filepath.Join(tempDir, "foo.txt"), []byte("meow"), 0644)
 			if err != nil {
 				t.Errorf("couldn't write tempfile %v", err)
 				t.FailNow()
@@ -144,21 +180,18 @@ func Test_CachingCopyCommand_ExecuteCommand(t *testing.T) {
 				},
 				fileContext: util.FileContext{Root: tempDir},
 				cmd: &instructions.CopyCommand{
-					SourcesAndDest: []string{
-						"foo.txt", "foo.txt",
-					},
-				},
+					SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{"foo.txt"}, DestPath: ""}},
 			}
 			count := 0
 			tc := testCase{
-				desctiption:    "with valid image and valid layer",
+				description:    "with valid image and valid layer",
 				count:          &count,
 				expectedCount:  1,
 				expectLayer:    true,
 				extractedFiles: []string{"/foo.txt"},
 				contextFiles:   []string{"foo.txt"},
 			}
-			c.extractFn = func(_ string, _ *tar.Header, _ io.Reader) error {
+			c.extractFn = func(_ string, _ *tar.Header, _ string, _ io.Reader) error {
 				*tc.count++
 				return nil
 			}
@@ -168,10 +201,10 @@ func Test_CachingCopyCommand_ExecuteCommand(t *testing.T) {
 		func() testCase {
 			c := &CachingCopyCommand{}
 			tc := testCase{
-				desctiption: "with no image",
+				description: "with no image",
 				expectErr:   true,
 			}
-			c.extractFn = func(_ string, _ *tar.Header, _ io.Reader) error {
+			c.extractFn = func(_ string, _ *tar.Header, _ string, _ io.Reader) error {
 				return nil
 			}
 			tc.command = c
@@ -181,11 +214,11 @@ func Test_CachingCopyCommand_ExecuteCommand(t *testing.T) {
 			c := &CachingCopyCommand{
 				img: fakeImage{},
 			}
-			c.extractFn = func(_ string, _ *tar.Header, _ io.Reader) error {
+			c.extractFn = func(_ string, _ *tar.Header, _ string, _ io.Reader) error {
 				return nil
 			}
 			return testCase{
-				desctiption: "with image containing no layers",
+				description: "with image containing no layers",
 				expectErr:   true,
 				command:     c,
 			}
@@ -198,11 +231,11 @@ func Test_CachingCopyCommand_ExecuteCommand(t *testing.T) {
 					},
 				},
 			}
-			c.extractFn = func(_ string, _ *tar.Header, _ io.Reader) error {
+			c.extractFn = func(_ string, _ *tar.Header, _ string, _ io.Reader) error {
 				return nil
 			}
 			tc := testCase{
-				desctiption: "with image one layer which has no tar content",
+				description: "with image one layer which has no tar content",
 				expectErr:   false, // this one probably should fail but doesn't because of how ExecuteCommand and util.GetFSFromLayers are implemented - cvgw- 2019-11-25
 				expectLayer: true,
 			}
@@ -212,7 +245,7 @@ func Test_CachingCopyCommand_ExecuteCommand(t *testing.T) {
 	}
 
 	for _, tc := range testCases {
-		t.Run(tc.desctiption, func(t *testing.T) {
+		t.Run(tc.description, func(t *testing.T) {
 			c := tc.command
 			err := c.ExecuteCommand(config, buildArgs)
 			if !tc.expectErr && err != nil {
@@ -261,8 +294,7 @@ func Test_CachingCopyCommand_ExecuteCommand(t *testing.T) {
 }
 
 func TestCopyExecuteCmd(t *testing.T) {
-	tempDir := setupTestTemp()
-	defer os.RemoveAll(tempDir)
+	tempDir := setupTestTemp(t)
 
 	cfg := &v1.Config{
 		Cmd:        nil,
@@ -277,7 +309,8 @@ func TestCopyExecuteCmd(t *testing.T) {
 
 			cmd := CopyCommand{
 				cmd: &instructions.CopyCommand{
-					SourcesAndDest: test.sourcesAndDest,
+					SourcesAndDest: instructions.SourcesAndDest{SourcePaths: test.sourcesAndDest[0 : len(test.sourcesAndDest)-1],
+						DestPath: test.sourcesAndDest[len(test.sourcesAndDest)-1]},
 				},
 				fileContext: fileContext,
 			}
@@ -299,13 +332,17 @@ func TestCopyExecuteCmd(t *testing.T) {
 			if err != nil {
 				t.Error()
 			}
+			if fstat == nil {
+				t.Error()
+				return // Unrecoverable, will segfault in the next line
+			}
 			if fstat.IsDir() {
-				files, err := ioutil.ReadDir(dest)
+				files, err := readDirectory(dest)
 				if err != nil {
 					t.Error()
 				}
 				for _, file := range files {
-					logrus.Debugf("file: %v", file.Name())
+					logrus.Debugf("File: %v", file.Name())
 					dirList = append(dirList, file.Name())
 				}
 			} else {
@@ -336,17 +373,14 @@ func Test_resolveIfSymlink(t *testing.T) {
 		err          error
 	}
 
-	tmpDir, err := ioutil.TempDir("", "copy-test")
+	tmpDir := t.TempDir()
+
+	baseDir, err := os.MkdirTemp(tmpDir, "not-linked")
 	if err != nil {
 		t.Error(err)
 	}
 
-	baseDir, err := ioutil.TempDir(tmpDir, "not-linked")
-	if err != nil {
-		t.Error(err)
-	}
-
-	path, err := ioutil.TempFile(baseDir, "foo.txt")
+	path, err := os.CreateTemp(baseDir, "foo.txt")
 	if err != nil {
 		t.Error(err)
 	}
@@ -381,7 +415,7 @@ func Test_resolveIfSymlink(t *testing.T) {
 	for i, c := range cases {
 		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
 			res, e := resolveIfSymlink(c.destPath)
-			if e != c.err {
+			if !errors.Is(e, c.err) {
 				t.Errorf("%s: expected %v but got %v", c.destPath, c.err, e)
 			}
 
@@ -392,12 +426,9 @@ func Test_resolveIfSymlink(t *testing.T) {
 	}
 }
 
-func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
+func Test_CopyEnvAndWildcards(t *testing.T) {
 	setupDirs := func(t *testing.T) (string, string) {
-		testDir, err := ioutil.TempDir("", "")
-		if err != nil {
-			t.Fatal(err)
-		}
+		testDir := t.TempDir()
 
 		dir := filepath.Join(testDir, "bar")
 
@@ -406,11 +437,114 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		}
 		file := filepath.Join(dir, "bam.txt")
 
-		if err := ioutil.WriteFile(file, []byte("meow"), 0777); err != nil {
+		if err := os.WriteFile(file, []byte("meow"), 0777); err != nil {
 			t.Fatal(err)
 		}
 		targetPath := filepath.Join(dir, "dam.txt")
-		if err := ioutil.WriteFile(targetPath, []byte("woof"), 0777); err != nil {
+		if err := os.WriteFile(targetPath, []byte("woof"), 0777); err != nil {
+			t.Fatal(err)
+		}
+		if err := os.Symlink("dam.txt", filepath.Join(dir, "sym.link")); err != nil {
+			t.Fatal(err)
+		}
+
+		return testDir, filepath.Base(dir)
+	}
+
+	t.Run("copy sources into a dir defined in env variable", func(t *testing.T) {
+		testDir, srcDir := setupDirs(t)
+		defer os.RemoveAll(testDir)
+		expected, err := readDirectory(filepath.Join(testDir, srcDir))
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		targetPath := filepath.Join(testDir, "target") + "/"
+
+		cmd := CopyCommand{
+			cmd: &instructions.CopyCommand{
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{srcDir + "/*"}, DestPath: "$TARGET_PATH"},
+			},
+			fileContext: util.FileContext{Root: testDir},
+		}
+
+		cfg := &v1.Config{
+			Cmd:        nil,
+			Env:        []string{"TARGET_PATH=" + targetPath},
+			WorkingDir: testDir,
+		}
+
+		err = cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
+		if err != nil {
+			t.Fatal(err)
+		}
+		testutil.CheckNoError(t, err)
+
+		actual, err := readDirectory(targetPath)
+		if err != nil {
+			t.Fatal(err)
+		}
+		for i, f := range actual {
+			testutil.CheckDeepEqual(t, expected[i].Name(), f.Name())
+			testutil.CheckDeepEqual(t, expected[i].Mode(), f.Mode())
+		}
+
+	})
+
+	t.Run("copy sources into a dir defined in env variable with no file found", func(t *testing.T) {
+		testDir, srcDir := setupDirs(t)
+		defer os.RemoveAll(testDir)
+
+		targetPath := filepath.Join(testDir, "target") + "/"
+
+		cmd := CopyCommand{
+			cmd: &instructions.CopyCommand{
+				//should only dam and bam be copied
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{srcDir + "/tam[s]"}, DestPath: "$TARGET_PATH"},
+			},
+			fileContext: util.FileContext{Root: testDir},
+		}
+
+		cfg := &v1.Config{
+			Cmd:        nil,
+			Env:        []string{"TARGET_PATH=" + targetPath},
+			WorkingDir: testDir,
+		}
+
+		err := cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
+		if err != nil {
+			t.Fatal(err)
+		}
+		testutil.CheckNoError(t, err)
+
+		actual, err := readDirectory(targetPath)
+
+		//check it should error out since no files are copied and targetPath is not created
+		if err == nil {
+			t.Fatal("expected error no dirrectory but got nil")
+		}
+
+		//actual should empty since no files are copied
+		testutil.CheckDeepEqual(t, 0, len(actual))
+	})
+}
+
+func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
+	setupDirs := func(t *testing.T) (string, string) {
+		testDir := t.TempDir()
+
+		dir := filepath.Join(testDir, "bar")
+
+		if err := os.MkdirAll(dir, 0777); err != nil {
+			t.Fatal(err)
+		}
+		file := filepath.Join(dir, "bam.txt")
+
+		if err := os.WriteFile(file, []byte("meow"), 0777); err != nil {
+			t.Fatal(err)
+		}
+		targetPath := filepath.Join(dir, "dam.txt")
+		if err := os.WriteFile(targetPath, []byte("woof"), 0777); err != nil {
 			t.Fatal(err)
 		}
 		if err := os.Symlink("dam.txt", filepath.Join(dir, "sym.link")); err != nil {
@@ -423,14 +557,14 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 	t.Run("copy dir to another dir", func(t *testing.T) {
 		testDir, srcDir := setupDirs(t)
 		defer os.RemoveAll(testDir)
-		expected, err := ioutil.ReadDir(filepath.Join(testDir, srcDir))
+		expected, err := readDirectory(filepath.Join(testDir, srcDir))
 		if err != nil {
 			t.Fatal(err)
 		}
 
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{srcDir, "dest"},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{srcDir}, DestPath: "dest"},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -447,7 +581,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		}
 		testutil.CheckNoError(t, err)
 		// Check if "dest" dir exists with contents of srcDir
-		actual, err := ioutil.ReadDir(filepath.Join(testDir, "dest"))
+		actual, err := readDirectory(filepath.Join(testDir, "dest"))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -457,12 +591,66 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		}
 	})
 
+	t.Run("copy dir to another dir - with ignored files", func(t *testing.T) {
+		testDir, srcDir := setupDirs(t)
+		defer os.RemoveAll(testDir)
+		ignoredFile := "bam.txt"
+		srcFiles, err := readDirectory(filepath.Join(testDir, srcDir))
+		if err != nil {
+			t.Fatal(err)
+		}
+		expected := map[string]fs.FileInfo{}
+		for _, sf := range srcFiles {
+			if sf.Name() == ignoredFile {
+				continue
+			}
+			expected[sf.Name()] = sf
+		}
+
+		cmd := CopyCommand{
+			cmd: &instructions.CopyCommand{
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{srcDir}, DestPath: "dest"},
+			},
+			fileContext: util.FileContext{
+				Root:          testDir,
+				ExcludedFiles: []string{filepath.Join(srcDir, ignoredFile)}},
+		}
+
+		cfg := &v1.Config{
+			Cmd:        nil,
+			Env:        []string{},
+			WorkingDir: testDir,
+		}
+
+		err = cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
+		if err != nil {
+			t.Fatal(err)
+		}
+		testutil.CheckNoError(t, err)
+		// Check if "dest" dir exists with contents of srcDir
+		actual, err := readDirectory(filepath.Join(testDir, "dest"))
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		if len(actual) != len(expected) {
+			t.Errorf("%v files are expected to be copied, but got %v", len(expected), len(actual))
+		}
+		for _, f := range actual {
+			if f.Name() == ignoredFile {
+				t.Errorf("file %v is expected to be ignored, but copied", f.Name())
+			}
+			testutil.CheckDeepEqual(t, expected[f.Name()].Name(), f.Name())
+			testutil.CheckDeepEqual(t, expected[f.Name()].Mode(), f.Mode())
+		}
+	})
+
 	t.Run("copy file to a dir", func(t *testing.T) {
 		testDir, srcDir := setupDirs(t)
 		defer os.RemoveAll(testDir)
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{filepath.Join(srcDir, "bam.txt"), "dest/"},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{filepath.Join(srcDir, "bam.txt")}, DestPath: "dest/"},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -476,7 +664,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		err := cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
 		testutil.CheckNoError(t, err)
 		// Check if "dest" dir exists with file bam.txt
-		files, err := ioutil.ReadDir(filepath.Join(testDir, "dest"))
+		files, err := readDirectory(filepath.Join(testDir, "dest"))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -489,7 +677,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		defer os.RemoveAll(testDir)
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{filepath.Join(srcDir, "bam.txt"), "dest"},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{filepath.Join(srcDir, "bam.txt")}, DestPath: "dest"},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -518,7 +706,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{filepath.Join(srcDir, "bam.txt"), "dest"},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{filepath.Join(srcDir, "bam.txt")}, DestPath: "dest"},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -532,7 +720,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		err := cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
 		testutil.CheckNoError(t, err)
 		// Check if "dest" dir exists with file bam.txt
-		files, err := ioutil.ReadDir(filepath.Join(testDir, "dest"))
+		files, err := readDirectory(filepath.Join(testDir, "dest"))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -547,7 +735,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{filepath.Join(srcDir, "sym.link"), "dest/"},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{filepath.Join(srcDir, "sym.link")}, DestPath: "dest/"},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -561,7 +749,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		err := cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
 		testutil.CheckNoError(t, err)
 		// Check if "dest" dir exists with link sym.link
-		files, err := ioutil.ReadDir(filepath.Join(testDir, "dest"))
+		files, err := readDirectory(filepath.Join(testDir, "dest"))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -580,7 +768,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		testDir, srcDir := setupDirs(t)
 		defer os.RemoveAll(testDir)
 		doesNotExists := filepath.Join(testDir, "dead.txt")
-		if err := ioutil.WriteFile(doesNotExists, []byte("remove me"), 0777); err != nil {
+		if err := os.WriteFile(doesNotExists, []byte("remove me"), 0777); err != nil {
 			t.Fatal(err)
 		}
 		if err := os.Symlink("../dead.txt", filepath.Join(testDir, srcDir, "dead.link")); err != nil {
@@ -592,7 +780,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{filepath.Join(srcDir, "dead.link"), "dest/"},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{filepath.Join(srcDir, "dead.link")}, DestPath: "dest/"},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -606,7 +794,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		err := cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
 		testutil.CheckNoError(t, err)
 		// Check if "dest" dir exists with link dead.link
-		files, err := ioutil.ReadDir(filepath.Join(testDir, "dest"))
+		files, err := readDirectory(filepath.Join(testDir, "dest"))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -623,7 +811,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 	t.Run("copy src symlink dir to a dir", func(t *testing.T) {
 		testDir, srcDir := setupDirs(t)
 		defer os.RemoveAll(testDir)
-		expected, err := ioutil.ReadDir(filepath.Join(testDir, srcDir))
+		expected, err := os.ReadDir(filepath.Join(testDir, srcDir))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -633,7 +821,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{"another", "dest"},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{"another"}, DestPath: "dest"},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -647,20 +835,20 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		err = cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
 		testutil.CheckNoError(t, err)
 		// Check if "dest" dir exists with contents of srcDir
-		actual, err := ioutil.ReadDir(filepath.Join(testDir, "dest"))
+		actual, err := os.ReadDir(filepath.Join(testDir, "dest"))
 		if err != nil {
 			t.Fatal(err)
 		}
 		for i, f := range actual {
 			testutil.CheckDeepEqual(t, expected[i].Name(), f.Name())
-			testutil.CheckDeepEqual(t, expected[i].Mode(), f.Mode())
+			testutil.CheckDeepEqual(t, expected[i].Type(), f.Type())
 		}
 	})
 
 	t.Run("copy dir with a symlink to a file outside of current src dir", func(t *testing.T) {
 		testDir, srcDir := setupDirs(t)
 		defer os.RemoveAll(testDir)
-		expected, err := ioutil.ReadDir(filepath.Join(testDir, srcDir))
+		expected, err := readDirectory(filepath.Join(testDir, srcDir))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -670,7 +858,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 			t.Fatal(err)
 		}
 		targetPath := filepath.Join(anotherSrc, "target.txt")
-		if err := ioutil.WriteFile(targetPath, []byte("woof"), 0777); err != nil {
+		if err := os.WriteFile(targetPath, []byte("woof"), 0777); err != nil {
 			t.Fatal(err)
 		}
 		if err := os.Symlink(targetPath, filepath.Join(testDir, srcDir, "zSym.link")); err != nil {
@@ -679,7 +867,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{srcDir, "dest"},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{srcDir}, DestPath: "dest"},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -694,7 +882,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		testutil.CheckNoError(t, err)
 		// Check if "dest" dir exists contents of srcDir and an extra zSym.link created
 		// in this test
-		actual, err := ioutil.ReadDir(filepath.Join(testDir, "dest"))
+		actual, err := readDirectory(filepath.Join(testDir, "dest"))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -713,7 +901,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 	t.Run("copy src symlink dir to a dir", func(t *testing.T) {
 		testDir, srcDir := setupDirs(t)
 		defer os.RemoveAll(testDir)
-		expected, err := ioutil.ReadDir(filepath.Join(testDir, srcDir))
+		expected, err := os.ReadDir(filepath.Join(testDir, srcDir))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -723,7 +911,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{"another", "dest"},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{"another"}, DestPath: "dest"},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -737,20 +925,20 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		err = cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
 		testutil.CheckNoError(t, err)
 		// Check if "dest" dir exists with bam.txt and "dest" dir is a symlink
-		actual, err := ioutil.ReadDir(filepath.Join(testDir, "dest"))
+		actual, err := os.ReadDir(filepath.Join(testDir, "dest"))
 		if err != nil {
 			t.Fatal(err)
 		}
 		for i, f := range actual {
 			testutil.CheckDeepEqual(t, expected[i].Name(), f.Name())
-			testutil.CheckDeepEqual(t, expected[i].Mode(), f.Mode())
+			testutil.CheckDeepEqual(t, expected[i].Type(), f.Type())
 		}
 	})
 
 	t.Run("copy src dir to a dest dir which is a symlink", func(t *testing.T) {
 		testDir, srcDir := setupDirs(t)
 		defer os.RemoveAll(testDir)
-		expected, err := ioutil.ReadDir(filepath.Join(testDir, srcDir))
+		expected, err := readDirectory(filepath.Join(testDir, srcDir))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -766,7 +954,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{srcDir, linkedDest},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{srcDir}, DestPath: linkedDest},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -780,7 +968,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		err = cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
 		testutil.CheckNoError(t, err)
 		// Check if "linkdest" dir exists with contents of srcDir
-		actual, err := ioutil.ReadDir(filepath.Join(testDir, "linkDest"))
+		actual, err := readDirectory(filepath.Join(testDir, "linkDest"))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -811,7 +999,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{fmt.Sprintf("%s/bam.txt", srcDir), linkedDest},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{fmt.Sprintf("%s/bam.txt", srcDir)}, DestPath: linkedDest},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -825,12 +1013,12 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		err := cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
 		testutil.CheckNoError(t, err)
 		// Check if "linkDest" link is same.
-		actual, err := ioutil.ReadDir(filepath.Join(testDir, "dest"))
+		actual, err := readDirectory(filepath.Join(testDir, "dest"))
 		if err != nil {
 			t.Fatal(err)
 		}
 		testutil.CheckDeepEqual(t, "bam.txt", actual[0].Name())
-		c, err := ioutil.ReadFile(filepath.Join(testDir, "dest", "bam.txt"))
+		c, err := os.ReadFile(filepath.Join(testDir, "dest", "bam.txt"))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -859,7 +1047,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{fmt.Sprintf("%s/bam.txt", srcDir), testDir},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{fmt.Sprintf("%s/bam.txt", srcDir)}, DestPath: testDir},
 				Chown:          "alice:group",
 			},
 			fileContext: util.FileContext{Root: testDir},
@@ -874,7 +1062,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		err := cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
 		testutil.CheckNoError(t, err)
 
-		actual, err := ioutil.ReadDir(filepath.Join(testDir))
+		actual, err := readDirectory(filepath.Join(testDir))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -904,7 +1092,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{fmt.Sprintf("%s/bam.txt", srcDir), testDir},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{fmt.Sprintf("%s/bam.txt", srcDir)}, DestPath: testDir},
 				Chown:          "missing:missing",
 			},
 			fileContext: util.FileContext{Root: testDir},
@@ -936,7 +1124,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		dest := filepath.Join(testDir, "copy")
 		cmd := CopyCommand{
 			cmd: &instructions.CopyCommand{
-				SourcesAndDest: []string{srcDir, dest},
+				SourcesAndDest: instructions.SourcesAndDest{SourcePaths: []string{srcDir}, DestPath: dest},
 			},
 			fileContext: util.FileContext{Root: testDir},
 		}
@@ -948,7 +1136,7 @@ func TestCopyCommand_ExecuteCommand_Extended(t *testing.T) {
 		}
 		err := cmd.ExecuteCommand(cfg, dockerfile.NewBuildArgs([]string{}))
 		testutil.CheckNoError(t, err)
-		actual, err := ioutil.ReadDir(filepath.Join(dest, "another"))
+		actual, err := os.ReadDir(filepath.Join(dest, "another"))
 		if err != nil {
 			t.Fatal(err)
 		}
