@@ -20,7 +20,6 @@ import (
 	"archive/tar"
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"reflect"
@@ -28,16 +27,19 @@ import (
 	"strconv"
 	"testing"
 
+	"github.com/GoogleContainerTools/kaniko/pkg/cache"
 	"github.com/GoogleContainerTools/kaniko/pkg/commands"
 	"github.com/GoogleContainerTools/kaniko/pkg/config"
 	"github.com/GoogleContainerTools/kaniko/pkg/dockerfile"
 	"github.com/GoogleContainerTools/kaniko/pkg/util"
 	"github.com/GoogleContainerTools/kaniko/testutil"
+	"github.com/containerd/containerd/platforms"
 	"github.com/google/go-cmp/cmp"
 	v1 "github.com/google/go-containerregistry/pkg/v1"
 	"github.com/google/go-containerregistry/pkg/v1/empty"
 	"github.com/google/go-containerregistry/pkg/v1/mutate"
 	"github.com/google/go-containerregistry/pkg/v1/partial"
+	"github.com/google/go-containerregistry/pkg/v1/types"
 	"github.com/moby/buildkit/frontend/dockerfile/instructions"
 )
 
@@ -367,10 +369,11 @@ COPY --from=second /bar /bat
 				initializeConfig = tt.args.mockInitConfig
 			}
 
-			f, _ := ioutil.TempFile("", "")
-			ioutil.WriteFile(f.Name(), []byte(tt.args.dockerfile), 0755)
+			f, _ := os.CreateTemp("", "")
+			os.WriteFile(f.Name(), []byte(tt.args.dockerfile), 0755)
 			opts := &config.KanikoOptions{
 				DockerfilePath: f.Name(),
+				CustomPlatform: platforms.Format(platforms.Normalize(platforms.DefaultSpec())),
 			}
 			testStages, metaArgs, err := dockerfile.ParseStages(opts)
 			if err != nil {
@@ -430,15 +433,11 @@ func Test_filesToSave(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			tmpDir, err := ioutil.TempDir("", "")
+			tmpDir := t.TempDir()
 			original := config.RootDir
 			config.RootDir = tmpDir
-			if err != nil {
-				t.Errorf("error creating tmpdir: %s", err)
-			}
 			defer func() {
 				config.RootDir = original
-				os.RemoveAll(tmpDir)
 			}()
 
 			for _, f := range tt.files {
@@ -464,6 +463,41 @@ func Test_filesToSave(t *testing.T) {
 			sort.Strings(got)
 			if !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("filesToSave() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestDeduplicatePaths(t *testing.T) {
+	tests := []struct {
+		name  string
+		input []string
+		want  []string
+	}{
+		{
+			name:  "no duplicates",
+			input: []string{"file1.txt", "file2.txt", "usr/lib"},
+			want:  []string{"file1.txt", "file2.txt", "usr/lib"},
+		},
+		{
+			name:  "duplicates",
+			input: []string{"file1.txt", "file2.txt", "file2.txt", "usr/lib"},
+			want:  []string{"file1.txt", "file2.txt", "usr/lib"},
+		},
+		{
+			name:  "duplicates with paths",
+			input: []string{"file1.txt", "file2.txt", "file2.txt", "usr/lib", "usr/lib/ssl"},
+			want:  []string{"file1.txt", "file2.txt", "usr/lib"},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := deduplicatePaths(tt.input)
+			sort.Strings(tt.want)
+			sort.Strings(got)
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("TestDeduplicatePaths() = %v, want %v", got, tt.want)
 			}
 		})
 	}
@@ -524,6 +558,36 @@ func TestInitializeConfig(t *testing.T) {
 	}
 }
 
+func Test_newLayerCache_defaultCache(t *testing.T) {
+	t.Run("default layer cache is registry cache", func(t *testing.T) {
+		layerCache := newLayerCache(&config.KanikoOptions{CacheRepo: "some-cache-repo"})
+		foundCache, ok := layerCache.(*cache.RegistryCache)
+		if !ok {
+			t.Error("expected layer cache to be a registry cache")
+		}
+		if foundCache.Opts.CacheRepo != "some-cache-repo" {
+			t.Errorf(
+				"expected cache repo to be 'some-cache-repo'; got %q", foundCache.Opts.CacheRepo,
+			)
+		}
+	})
+}
+
+func Test_newLayerCache_layoutCache(t *testing.T) {
+	t.Run("when cache repo has 'oci:' prefix layer cache is layout cache", func(t *testing.T) {
+		layerCache := newLayerCache(&config.KanikoOptions{CacheRepo: "oci:/some-cache-repo"})
+		foundCache, ok := layerCache.(*cache.LayoutCache)
+		if !ok {
+			t.Error("expected layer cache to be a layout cache")
+		}
+		if foundCache.Opts.CacheRepo != "oci:/some-cache-repo" {
+			t.Errorf(
+				"expected cache repo to be 'oci:/some-cache-repo'; got %q", foundCache.Opts.CacheRepo,
+			)
+		}
+	})
+}
+
 func Test_stageBuilder_optimize(t *testing.T) {
 	testCases := []struct {
 		opts     *config.KanikoOptions
@@ -552,12 +616,12 @@ func Test_stageBuilder_optimize(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			cf := &v1.ConfigFile{}
-			snap := fakeSnapShotter{}
+			snap := &fakeSnapShotter{}
 			lc := &fakeLayerCache{retrieve: tc.retrieve}
 			sb := &stageBuilder{opts: tc.opts, cf: cf, snapshotter: snap, layerCache: lc,
 				args: dockerfile.NewBuildArgs([]string{})}
 			ck := CompositeCache{}
-			file, err := ioutil.TempFile("", "foo")
+			file, err := os.CreateTemp("", "foo")
 			if err != nil {
 				t.Error(err)
 			}
@@ -590,14 +654,15 @@ func newStageContext(command string, args map[string]string, env []string) stage
 }
 
 func Test_stageBuilder_populateCompositeKey(t *testing.T) {
-	testCases := []struct {
+	type testcase struct {
 		description string
 		cmd1        stageContext
 		cmd2        stageContext
 		shdEqual    bool
-	}{
+	}
+	testCases := []testcase{
 		{
-			description: "cache key for same command with same build args",
+			description: "cache key for same command [RUN] with same build args",
 			cmd1: newStageContext(
 				"RUN echo $ARG > test",
 				map[string]string{"ARG": "foo"},
@@ -611,7 +676,7 @@ func Test_stageBuilder_populateCompositeKey(t *testing.T) {
 			shdEqual: true,
 		},
 		{
-			description: "cache key for same command with same env and args",
+			description: "cache key for same command [RUN] with same env and args",
 			cmd1: newStageContext(
 				"RUN echo $ENV > test",
 				map[string]string{"ARG": "foo"},
@@ -625,7 +690,7 @@ func Test_stageBuilder_populateCompositeKey(t *testing.T) {
 			shdEqual: true,
 		},
 		{
-			description: "cache key for same command with same env but different args",
+			description: "cache key for same command [RUN] with same env but different args",
 			cmd1: newStageContext(
 				"RUN echo $ENV > test",
 				map[string]string{"ARG": "foo"},
@@ -638,7 +703,7 @@ func Test_stageBuilder_populateCompositeKey(t *testing.T) {
 			),
 		},
 		{
-			description: "cache key for same command, different buildargs, args not used in command",
+			description: "cache key for same command [RUN], different buildargs, args not used in command",
 			cmd1: newStageContext(
 				"RUN echo const > test",
 				map[string]string{"ARG": "foo"},
@@ -651,7 +716,7 @@ func Test_stageBuilder_populateCompositeKey(t *testing.T) {
 			),
 		},
 		{
-			description: "cache key for same command, different buildargs, args used in script",
+			description: "cache key for same command [RUN], different buildargs, args used in script",
 			// test.sh
 			// #!/bin/sh
 			// echo ${ARG}
@@ -667,7 +732,7 @@ func Test_stageBuilder_populateCompositeKey(t *testing.T) {
 			),
 		},
 		{
-			description: "cache key for same command with a build arg values",
+			description: "cache key for same command [RUN] with a build arg values",
 			cmd1: newStageContext(
 				"RUN echo $ARG > test",
 				map[string]string{"ARG": "foo"},
@@ -680,7 +745,7 @@ func Test_stageBuilder_populateCompositeKey(t *testing.T) {
 			),
 		},
 		{
-			description: "cache key for same command with different env values",
+			description: "cache key for same command [RUN] with different env values",
 			cmd1: newStageContext(
 				"RUN echo $ENV > test",
 				map[string]string{"ARG": "foo"},
@@ -693,7 +758,7 @@ func Test_stageBuilder_populateCompositeKey(t *testing.T) {
 			),
 		},
 		{
-			description: "cache key for different command same context",
+			description: "cache key for different command [RUN] same context",
 			cmd1: newStageContext(
 				"RUN echo other > test",
 				map[string]string{"ARG": "foo"},
@@ -705,17 +770,133 @@ func Test_stageBuilder_populateCompositeKey(t *testing.T) {
 				[]string{"ENV=1"},
 			),
 		},
+		{
+			description: "cache key for command [RUN] with same env values [check that variable no interpolate in RUN command]",
+			cmd1: newStageContext(
+				"RUN echo $ENV > test",
+				map[string]string{"ARG": "foo"},
+				[]string{"ENV=1"},
+			),
+			cmd2: newStageContext(
+				"RUN echo 1 > test",
+				map[string]string{"ARG": "foo"},
+				[]string{"ENV=1"},
+			),
+			shdEqual: false,
+		},
+		{
+			description: "cache key for command [RUN] with different env values [check that variable no interpolate in RUN command]",
+			cmd1: newStageContext(
+				"RUN echo ${APP_VERSION%.*} ${APP_VERSION%-*} > test",
+				map[string]string{"ARG": "foo"},
+				[]string{"ENV=1"},
+			),
+			cmd2: newStageContext(
+				"RUN echo ${APP_VERSION%.*} ${APP_VERSION%-*} > test",
+				map[string]string{"ARG": "foo"},
+				[]string{"ENV=2"},
+			),
+			shdEqual: false,
+		},
+		func() testcase {
+			dir, files := tempDirAndFile(t)
+			file := files[0]
+			filePath := filepath.Join(dir, file)
+			return testcase{
+				description: "cache key for same command [COPY] with same args",
+				cmd1: newStageContext(
+					fmt.Sprintf("COPY %s /meow", filePath),
+					map[string]string{"ARG": "foo"},
+					[]string{"ENV=1"},
+				),
+				cmd2: newStageContext(
+					fmt.Sprintf("COPY %s /meow", filePath),
+					map[string]string{"ARG": "foo"},
+					[]string{"ENV=1"},
+				),
+				shdEqual: true,
+			}
+		}(),
+		func() testcase {
+			dir, files := tempDirAndFile(t)
+			file := files[0]
+			filePath := filepath.Join(dir, file)
+			return testcase{
+				description: "cache key for same command [COPY] with different args",
+				cmd1: newStageContext(
+					fmt.Sprintf("COPY %s /meow", filePath),
+					map[string]string{"ARG": "foo"},
+					[]string{"ENV=1"},
+				),
+				cmd2: newStageContext(
+					fmt.Sprintf("COPY %s /meow", filePath),
+					map[string]string{"ARG": "bar"},
+					[]string{"ENV=2"},
+				),
+				shdEqual: true,
+			}
+		}(),
+		{
+			description: "cache key for same command [WORKDIR] with same args",
+			cmd1: newStageContext(
+				"WORKDIR /",
+				map[string]string{"ARG": "foo"},
+				[]string{"ENV=1"},
+			),
+			cmd2: newStageContext(
+				"WORKDIR /",
+				map[string]string{"ARG": "foo"},
+				[]string{"ENV=1"},
+			),
+			shdEqual: true,
+		},
+		{
+			description: "cache key for same command [WORKDIR] with different args",
+			cmd1: newStageContext(
+				"WORKDIR /",
+				map[string]string{"ARG": "foo"},
+				[]string{"ENV=1"},
+			),
+			cmd2: newStageContext(
+				"WORKDIR /",
+				map[string]string{"ARG": "bar"},
+				[]string{"ENV=2"},
+			),
+			shdEqual: true,
+		},
 	}
 	for _, tc := range testCases {
 		t.Run(tc.description, func(t *testing.T) {
 			sb := &stageBuilder{fileContext: util.FileContext{Root: "workspace"}}
 			ck := CompositeCache{}
 
-			ck1, err := sb.populateCompositeKey(tc.cmd1.command, []string{}, ck, tc.cmd1.args, tc.cmd1.env)
+			instructions1, err := dockerfile.ParseCommands([]string{tc.cmd1.command.String()})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			fc1 := util.FileContext{Root: "workspace"}
+			dockerCommand1, err := commands.GetCommand(instructions1[0], fc1, false, true, true)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			instructions, err := dockerfile.ParseCommands([]string{tc.cmd2.command.String()})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			fc2 := util.FileContext{Root: "workspace"}
+			dockerCommand2, err := commands.GetCommand(instructions[0], fc2, false, true, true)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			ck1, err := sb.populateCompositeKey(dockerCommand1, []string{}, ck, tc.cmd1.args, tc.cmd1.env)
 			if err != nil {
 				t.Errorf("Expected error to be nil but was %v", err)
 			}
-			ck2, err := sb.populateCompositeKey(tc.cmd2.command, []string{}, ck, tc.cmd2.args, tc.cmd2.env)
+			ck2, err := sb.populateCompositeKey(dockerCommand2, []string{}, ck, tc.cmd2.args, tc.cmd2.env)
 			if err != nil {
 				t.Errorf("Expected error to be nil but was %v", err)
 			}
@@ -729,17 +910,21 @@ func Test_stageBuilder_populateCompositeKey(t *testing.T) {
 
 func Test_stageBuilder_build(t *testing.T) {
 	type testcase struct {
-		description       string
-		opts              *config.KanikoOptions
-		args              map[string]string
-		layerCache        *fakeLayerCache
-		expectedCacheKeys []string
-		pushedCacheKeys   []string
-		commands          []commands.DockerCommand
-		fileName          string
-		rootDir           string
-		image             v1.Image
-		config            *v1.ConfigFile
+		description        string
+		opts               *config.KanikoOptions
+		args               map[string]string
+		layerCache         *fakeLayerCache
+		expectedCacheKeys  []string
+		pushedCacheKeys    []string
+		commands           []commands.DockerCommand
+		fileName           string
+		rootDir            string
+		image              v1.Image
+		config             *v1.ConfigFile
+		stage              config.KanikoStage
+		crossStageDeps     map[int][]string
+		mockGetFSFromImage func(root string, img v1.Image, extract util.ExtractFunction) ([]string, error)
+		shouldInitSnapshot bool
 	}
 
 	testCases := []testcase{
@@ -762,10 +947,7 @@ func Test_stageBuilder_build(t *testing.T) {
 				},
 			}
 
-			destDir, err := ioutil.TempDir("", "baz")
-			if err != nil {
-				t.Errorf("could not create temp dir %v", err)
-			}
+			destDir := t.TempDir()
 			return testcase{
 				description:       "fake command cache enabled but key not in cache",
 				config:            &v1.ConfigFile{Config: v1.Config{WorkingDir: destDir}},
@@ -795,10 +977,7 @@ func Test_stageBuilder_build(t *testing.T) {
 				},
 			}
 
-			destDir, err := ioutil.TempDir("", "baz")
-			if err != nil {
-				t.Errorf("could not create temp dir %v", err)
-			}
+			destDir := t.TempDir()
 			return testcase{
 				description: "fake command cache enabled and key in cache",
 				opts:        &config.KanikoOptions{Cache: true},
@@ -831,10 +1010,7 @@ func Test_stageBuilder_build(t *testing.T) {
 				},
 			}
 
-			destDir, err := ioutil.TempDir("", "baz")
-			if err != nil {
-				t.Errorf("could not create temp dir %v", err)
-			}
+			destDir := t.TempDir()
 			return testcase{
 				description: "fake command cache enabled with tar compression disabled and key in cache",
 				opts:        &config.KanikoOptions{Cache: true, CompressedCaching: false},
@@ -848,6 +1024,15 @@ func Test_stageBuilder_build(t *testing.T) {
 				rootDir:           dir,
 			}
 		}(),
+		{
+			description: "use new run",
+			opts:        &config.KanikoOptions{RunV2: true},
+		},
+		{
+			description:        "single snapshot",
+			opts:               &config.KanikoOptions{SingleSnapshot: true},
+			shouldInitSnapshot: true,
+		},
 		{
 			description: "fake command cache disabled and key not in cache",
 			opts:        &config.KanikoOptions{Cache: false},
@@ -875,17 +1060,16 @@ func Test_stageBuilder_build(t *testing.T) {
 			}
 			copyCommandCacheKey := hash
 			dockerFile := fmt.Sprintf(`
-FROM ubuntu:16.04
-COPY %s foo.txt
-`, filename)
-			f, _ := ioutil.TempFile("", "")
-			ioutil.WriteFile(f.Name(), []byte(dockerFile), 0755)
+		FROM ubuntu:16.04
+		COPY %s foo.txt
+		`, filename)
+			f, _ := os.CreateTemp("", "")
+			os.WriteFile(f.Name(), []byte(dockerFile), 0755)
 			opts := &config.KanikoOptions{
 				DockerfilePath:  f.Name(),
 				Cache:           true,
 				CacheCopyLayers: true,
 			}
-
 			testStages, metaArgs, err := dockerfile.ParseStages(opts)
 			if err != nil {
 				t.Errorf("Failed to parse test dockerfile to stages: %s", err)
@@ -924,7 +1108,7 @@ COPY %s foo.txt
 				expectedCacheKeys: []string{copyCommandCacheKey},
 				// CachingCopyCommand is not pushed to the cache
 				pushedCacheKeys: []string{},
-				commands:        getCommands(util.FileContext{Root: dir}, cmds, true),
+				commands:        getCommands(util.FileContext{Root: dir}, cmds, true, false),
 				fileName:        filename,
 			}
 		}(),
@@ -932,10 +1116,7 @@ COPY %s foo.txt
 			dir, filenames := tempDirAndFile(t)
 			filename := filenames[0]
 			tarContent := []byte{}
-			destDir, err := ioutil.TempDir("", "baz")
-			if err != nil {
-				t.Errorf("could not create temp dir %v", err)
-			}
+			destDir := t.TempDir()
 			filePath := filepath.Join(dir, filename)
 			ch := NewCompositeCache("", fmt.Sprintf("COPY %s foo.txt", filename))
 			ch.AddPath(filePath, util.FileContext{})
@@ -948,8 +1129,8 @@ COPY %s foo.txt
 FROM ubuntu:16.04
 COPY %s foo.txt
 `, filename)
-			f, _ := ioutil.TempFile("", "")
-			ioutil.WriteFile(f.Name(), []byte(dockerFile), 0755)
+			f, _ := os.CreateTemp("", "")
+			os.WriteFile(f.Name(), []byte(dockerFile), 0755)
 			opts := &config.KanikoOptions{
 				DockerfilePath:  f.Name(),
 				Cache:           true,
@@ -984,7 +1165,7 @@ COPY %s foo.txt
 				rootDir:           dir,
 				expectedCacheKeys: []string{hash},
 				pushedCacheKeys:   []string{hash},
-				commands:          getCommands(util.FileContext{Root: dir}, cmds, true),
+				commands:          getCommands(util.FileContext{Root: dir}, cmds, true, false),
 				fileName:          filename,
 			}
 		}(),
@@ -993,10 +1174,7 @@ COPY %s foo.txt
 			filename := filenames[0]
 			tarContent := generateTar(t, filename)
 
-			destDir, err := ioutil.TempDir("", "baz")
-			if err != nil {
-				t.Errorf("could not create temp dir %v", err)
-			}
+			destDir := t.TempDir()
 			filePath := filepath.Join(dir, filename)
 
 			ch := NewCompositeCache("", "RUN foobar")
@@ -1030,8 +1208,8 @@ FROM ubuntu:16.04
 RUN foobar
 COPY %s bar.txt
 `, filename)
-			f, _ := ioutil.TempFile("", "")
-			ioutil.WriteFile(f.Name(), []byte(dockerFile), 0755)
+			f, _ := os.CreateTemp("", "")
+			os.WriteFile(f.Name(), []byte(dockerFile), 0755)
 			opts := &config.KanikoOptions{
 				DockerfilePath: f.Name(),
 			}
@@ -1051,7 +1229,7 @@ COPY %s bar.txt
 			cmds := stage.Commands
 			return testcase{
 				description: "cached run command followed by uncached copy command results in consistent read and write hashes",
-				opts:        &config.KanikoOptions{Cache: true, CacheCopyLayers: true},
+				opts:        &config.KanikoOptions{Cache: true, CacheCopyLayers: true, CacheRunLayers: true},
 				rootDir:     dir,
 				config:      &v1.ConfigFile{Config: v1.Config{WorkingDir: destDir}},
 				layerCache: &fakeLayerCache{
@@ -1062,7 +1240,7 @@ COPY %s bar.txt
 				// hash1 is the read cachekey for the first layer
 				expectedCacheKeys: []string{hash1, hash2},
 				pushedCacheKeys:   []string{hash2},
-				commands:          getCommands(util.FileContext{Root: dir}, cmds, true),
+				commands:          getCommands(util.FileContext{Root: dir}, cmds, true, true),
 			}
 		}(),
 		func() testcase {
@@ -1070,10 +1248,7 @@ COPY %s bar.txt
 			filename := filenames[0]
 			tarContent := generateTar(t, filename)
 
-			destDir, err := ioutil.TempDir("", "baz")
-			if err != nil {
-				t.Errorf("could not create temp dir %v", err)
-			}
+			destDir := t.TempDir()
 
 			filePath := filepath.Join(dir, filename)
 
@@ -1081,12 +1256,12 @@ COPY %s bar.txt
 			ch.AddPath(filePath, util.FileContext{})
 
 			// copy hash
-			_, err = ch.Hash()
+			_, err := ch.Hash()
 			if err != nil {
 				t.Errorf("couldn't create hash %v", err)
 			}
 
-			ch.AddKey(fmt.Sprintf("RUN foobar"))
+			ch.AddKey("RUN foobar")
 
 			// run hash
 			runHash, err := ch.Hash()
@@ -1107,8 +1282,8 @@ FROM ubuntu:16.04
 COPY %s bar.txt
 RUN foobar
 `, filename)
-			f, _ := ioutil.TempFile("", "")
-			ioutil.WriteFile(f.Name(), []byte(dockerFile), 0755)
+			f, _ := os.CreateTemp("", "")
+			os.WriteFile(f.Name(), []byte(dockerFile), 0755)
 			opts := &config.KanikoOptions{
 				DockerfilePath: f.Name(),
 			}
@@ -1128,7 +1303,7 @@ RUN foobar
 			cmds := stage.Commands
 			return testcase{
 				description: "copy command followed by cached run command results in consistent read and write hashes",
-				opts:        &config.KanikoOptions{Cache: true},
+				opts:        &config.KanikoOptions{Cache: true, CacheRunLayers: true},
 				rootDir:     dir,
 				config:      &v1.ConfigFile{Config: v1.Config{WorkingDir: destDir}},
 				layerCache: &fakeLayerCache{
@@ -1138,7 +1313,7 @@ RUN foobar
 				image:             image,
 				expectedCacheKeys: []string{runHash},
 				pushedCacheKeys:   []string{},
-				commands:          getCommands(util.FileContext{Root: dir}, cmds, false),
+				commands:          getCommands(util.FileContext{Root: dir}, cmds, false, true),
 			}
 		}(),
 		func() testcase {
@@ -1158,6 +1333,7 @@ RUN foobar
 				cacheCommand: MockCachedDockerCommand{
 					contextFiles: []string{},
 				},
+				argToCompositeCache: true,
 			}
 
 			return testcase{
@@ -1183,7 +1359,7 @@ RUN foobar
 			ch := NewCompositeCache("")
 			ch.AddKey("|1")
 			ch.AddKey("arg=value")
-			ch.AddKey("RUN value")
+			ch.AddKey("RUN $arg")
 			hash, err := ch.Hash()
 			if err != nil {
 				t.Errorf("couldn't create hash %v", err)
@@ -1195,6 +1371,7 @@ RUN foobar
 				cacheCommand: MockCachedDockerCommand{
 					contextFiles: []string{},
 				},
+				argToCompositeCache: true,
 			}
 
 			return testcase{
@@ -1227,7 +1404,7 @@ RUN foobar
 			ch2 := NewCompositeCache("")
 			ch2.AddKey("|1")
 			ch2.AddKey("arg=anotherValue")
-			ch2.AddKey("RUN anotherValue")
+			ch2.AddKey("RUN $arg")
 			hash2, err := ch2.Hash()
 			if err != nil {
 				t.Errorf("couldn't create hash %v", err)
@@ -1238,6 +1415,7 @@ RUN foobar
 				cacheCommand: MockCachedDockerCommand{
 					contextFiles: []string{},
 				},
+				argToCompositeCache: true,
 			}
 
 			return testcase{
@@ -1258,12 +1436,21 @@ RUN foobar
 				rootDir:           dir,
 			}
 		}(),
+		{
+			description:    "fs unpacked",
+			opts:           &config.KanikoOptions{InitialFSUnpacked: true},
+			stage:          config.KanikoStage{Index: 0},
+			crossStageDeps: map[int][]string{0: {"some-dep"}},
+			mockGetFSFromImage: func(root string, img v1.Image, extract util.ExtractFunction) ([]string, error) {
+				return nil, fmt.Errorf("getFSFromImage shouldn't be called if fs is already unpacked")
+			},
+		},
 	}
 	for _, tc := range testCases {
 		t.Run(tc.description, func(t *testing.T) {
 			var fileName string
 			if tc.commands == nil {
-				file, err := ioutil.TempFile("", "foo")
+				file, err := os.CreateTemp("", "foo")
 				if err != nil {
 					t.Error(err)
 				}
@@ -1288,7 +1475,7 @@ RUN foobar
 				}
 			}
 
-			snap := fakeSnapShotter{file: fileName}
+			snap := &fakeSnapShotter{file: fileName}
 			lc := tc.layerCache
 			if lc == nil {
 				lc = &fakeLayerCache{}
@@ -1314,11 +1501,22 @@ RUN foobar
 			if tc.rootDir != "" {
 				config.RootDir = tc.rootDir
 			}
+			sb.stage = tc.stage
+			sb.crossStageDeps = tc.crossStageDeps
+			if tc.mockGetFSFromImage != nil {
+				original := getFSFromImage
+				defer func() { getFSFromImage = original }()
+				getFSFromImage = tc.mockGetFSFromImage
+			}
 			err := sb.build()
 			if err != nil {
 				t.Errorf("Expected error to be nil but was %v", err)
 			}
-			fmt.Println(lc.receivedKeys)
+			if tc.shouldInitSnapshot && !snap.initialized {
+				t.Errorf("Snapshotter was not initialized but should have been")
+			} else if !tc.shouldInitSnapshot && snap.initialized {
+				t.Errorf("Snapshotter was initialized but should not have been")
+			}
 			assertCacheKeys(t, tc.expectedCacheKeys, lc.receivedKeys, "receive")
 			assertCacheKeys(t, tc.pushedCacheKeys, keys, "push")
 
@@ -1351,7 +1549,7 @@ func assertCacheKeys(t *testing.T, expectedCacheKeys, actualCacheKeys []string, 
 	}
 }
 
-func getCommands(fileContext util.FileContext, cmds []instructions.Command, cacheCopy bool) []commands.DockerCommand {
+func getCommands(fileContext util.FileContext, cmds []instructions.Command, cacheCopy, cacheRun bool) []commands.DockerCommand {
 	outCommands := make([]commands.DockerCommand, 0)
 	for _, c := range cmds {
 		cmd, err := commands.GetCommand(
@@ -1359,6 +1557,7 @@ func getCommands(fileContext util.FileContext, cmds []instructions.Command, cach
 			fileContext,
 			false,
 			cacheCopy,
+			cacheRun,
 		)
 		if err != nil {
 			panic(err)
@@ -1366,19 +1565,15 @@ func getCommands(fileContext util.FileContext, cmds []instructions.Command, cach
 		outCommands = append(outCommands, cmd)
 	}
 	return outCommands
-
 }
 
 func tempDirAndFile(t *testing.T) (string, []string) {
 	filenames := []string{"bar.txt"}
 
-	dir, err := ioutil.TempDir("", "foo")
-	if err != nil {
-		t.Errorf("could not create temp dir %v", err)
-	}
+	dir := t.TempDir()
 	for _, filename := range filenames {
 		filepath := filepath.Join(dir, filename)
-		err = ioutil.WriteFile(filepath, []byte(`meow`), 0777)
+		err := os.WriteFile(filepath, []byte(`meow`), 0777)
 		if err != nil {
 			t.Errorf("could not create temp file %v", err)
 		}
@@ -1407,7 +1602,7 @@ func generateTar(t *testing.T, dir string, fileNames ...string) []byte {
 			t.Errorf("could not write tar header %v", err)
 		}
 
-		content, err := ioutil.ReadFile(filePath)
+		content, err := os.ReadFile(filePath)
 		if err != nil {
 			t.Errorf("could not read tempfile %v", err)
 		}
@@ -1440,9 +1635,10 @@ func Test_stageBuild_populateCompositeKeyForCopyCommand(t *testing.T) {
 		expectedCacheKey string
 	}{
 		{
-			description:      "multi-stage copy command",
+			description: "multi-stage copy command",
+			// dont use digest from previoust stage for COPY
 			command:          "COPY --from=0 foo.txt bar.txt",
-			expectedCacheKey: "COPY --from=0 foo.txt bar.txt-some-cache-key",
+			expectedCacheKey: "COPY --from=0 foo.txt bar.txt",
 		},
 		{
 			description:      "copy command",
@@ -1457,14 +1653,14 @@ func Test_stageBuild_populateCompositeKeyForCopyCommand(t *testing.T) {
 			}
 
 			fc := util.FileContext{Root: "workspace"}
-			copyCommand, err := commands.GetCommand(instructions[0], fc, false, true)
+			copyCommand, err := commands.GetCommand(instructions[0], fc, false, true, true)
 			if err != nil {
 				t.Fatal(err)
 			}
 
 			for _, useCacheCommand := range []bool{false, true} {
 				t.Run(fmt.Sprintf("CacheCommand=%t", useCacheCommand), func(t *testing.T) {
-					var cmd fmt.Stringer = copyCommand
+					var cmd commands.DockerCommand = copyCommand
 					if useCacheCommand {
 						cmd = copyCommand.(*commands.CopyCommand).CacheCommand(nil)
 					}
@@ -1547,5 +1743,282 @@ func Test_ResolveCrossStageInstructions(t *testing.T) {
 
 		expectedMap := map[string]string{"second": "1", "third": "2"}
 		testutil.CheckDeepEqual(t, expectedMap, stageToIdx)
+	}
+}
+
+func Test_stageBuilder_saveSnapshotToLayer(t *testing.T) {
+	dir, files := tempDirAndFile(t)
+	type fields struct {
+		stage            config.KanikoStage
+		image            v1.Image
+		cf               *v1.ConfigFile
+		baseImageDigest  string
+		finalCacheKey    string
+		opts             *config.KanikoOptions
+		fileContext      util.FileContext
+		cmds             []commands.DockerCommand
+		args             *dockerfile.BuildArgs
+		crossStageDeps   map[int][]string
+		digestToCacheKey map[string]string
+		stageIdxToDigest map[string]string
+		snapshotter      snapShotter
+		layerCache       cache.LayerCache
+		pushLayerToCache cachePusher
+	}
+	type args struct {
+		tarPath string
+	}
+	tests := []struct {
+		name              string
+		fields            fields
+		args              args
+		expectedMediaType types.MediaType
+		expectedDiff      v1.Hash
+		expectedDigest    v1.Hash
+		wantErr           bool
+	}{
+		{
+			name: "oci image",
+			fields: fields{
+				image: ociFakeImage{},
+				opts: &config.KanikoOptions{
+					ForceBuildMetadata: true,
+				},
+			},
+			args: args{
+				tarPath: filepath.Join(dir, files[0]),
+			},
+			expectedMediaType: types.OCILayer,
+			expectedDiff: v1.Hash{
+				Algorithm: "sha256",
+				Hex:       "404cdd7bc109c432f8cc2443b45bcfe95980f5107215c645236e577929ac3e52",
+			},
+			expectedDigest: v1.Hash{
+				Algorithm: "sha256",
+				Hex:       "1dc5887a31ec6b388646be46c5f0b2036f92f4cbba50d12163a8be4074565a91",
+			},
+		},
+		{
+			name: "docker image",
+			fields: fields{
+				image: fakeImage{},
+				opts: &config.KanikoOptions{
+					ForceBuildMetadata: true,
+				},
+			},
+			args: args{
+				tarPath: filepath.Join(dir, files[0]),
+			},
+			expectedMediaType: types.DockerLayer,
+			expectedDiff: v1.Hash{
+				Algorithm: "sha256",
+				Hex:       "404cdd7bc109c432f8cc2443b45bcfe95980f5107215c645236e577929ac3e52",
+			},
+			expectedDigest: v1.Hash{
+				Algorithm: "sha256",
+				Hex:       "1dc5887a31ec6b388646be46c5f0b2036f92f4cbba50d12163a8be4074565a91",
+			},
+		},
+		{
+			name: "oci image, zstd compression",
+			fields: fields{
+				image: ociFakeImage{},
+				opts: &config.KanikoOptions{
+					ForceBuildMetadata: true,
+					Compression:        config.ZStd,
+				},
+			},
+			args: args{
+				tarPath: filepath.Join(dir, files[0]),
+			},
+			expectedMediaType: types.OCILayerZStd,
+			expectedDiff: v1.Hash{
+				Algorithm: "sha256",
+				Hex:       "404cdd7bc109c432f8cc2443b45bcfe95980f5107215c645236e577929ac3e52",
+			},
+			expectedDigest: v1.Hash{
+				Algorithm: "sha256",
+				Hex:       "28369c11d9b68c9877781eaf4d8faffb4d0ada1900a1fb83ad452e58a072b45b",
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			s := &stageBuilder{
+				stage:            tt.fields.stage,
+				image:            tt.fields.image,
+				cf:               tt.fields.cf,
+				baseImageDigest:  tt.fields.baseImageDigest,
+				finalCacheKey:    tt.fields.finalCacheKey,
+				opts:             tt.fields.opts,
+				fileContext:      tt.fields.fileContext,
+				cmds:             tt.fields.cmds,
+				args:             tt.fields.args,
+				crossStageDeps:   tt.fields.crossStageDeps,
+				digestToCacheKey: tt.fields.digestToCacheKey,
+				stageIdxToDigest: tt.fields.stageIdxToDigest,
+				snapshotter:      tt.fields.snapshotter,
+				layerCache:       tt.fields.layerCache,
+				pushLayerToCache: tt.fields.pushLayerToCache,
+			}
+			got, err := s.saveSnapshotToLayer(tt.args.tarPath)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("stageBuilder.saveSnapshotToLayer() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if mt, _ := got.MediaType(); mt != tt.expectedMediaType {
+				t.Errorf("expected mediatype %s, got %s", tt.expectedMediaType, mt)
+				return
+			}
+			if diff, _ := got.DiffID(); diff != tt.expectedDiff {
+				t.Errorf("expected diff %s, got %s", tt.expectedDiff, diff)
+				return
+			}
+			if digest, _ := got.Digest(); digest != tt.expectedDigest {
+				t.Errorf("expected digest %s, got %s", tt.expectedDigest, digest)
+				return
+			}
+		})
+	}
+}
+
+func Test_stageBuilder_convertLayerMediaType(t *testing.T) {
+	type fields struct {
+		stage            config.KanikoStage
+		image            v1.Image
+		cf               *v1.ConfigFile
+		baseImageDigest  string
+		finalCacheKey    string
+		opts             *config.KanikoOptions
+		fileContext      util.FileContext
+		cmds             []commands.DockerCommand
+		args             *dockerfile.BuildArgs
+		crossStageDeps   map[int][]string
+		digestToCacheKey map[string]string
+		stageIdxToDigest map[string]string
+		snapshotter      snapShotter
+		layerCache       cache.LayerCache
+		pushLayerToCache cachePusher
+	}
+	type args struct {
+		layer v1.Layer
+	}
+	tests := []struct {
+		name              string
+		fields            fields
+		args              args
+		expectedMediaType types.MediaType
+		wantErr           bool
+	}{
+		{
+			name: "docker image w/ docker layer",
+			fields: fields{
+				image: fakeImage{},
+			},
+			args: args{
+				layer: fakeLayer{
+					mediaType: types.DockerLayer,
+				},
+			},
+			expectedMediaType: types.DockerLayer,
+		},
+		{
+			name: "oci image w/ oci layer",
+			fields: fields{
+				image: ociFakeImage{},
+			},
+			args: args{
+				layer: fakeLayer{
+					mediaType: types.OCILayer,
+				},
+			},
+			expectedMediaType: types.OCILayer,
+		},
+		{
+			name: "oci image w/ convertable docker layer",
+			fields: fields{
+				image: ociFakeImage{},
+				opts:  &config.KanikoOptions{},
+			},
+			args: args{
+				layer: fakeLayer{
+					mediaType: types.DockerLayer,
+				},
+			},
+			expectedMediaType: types.OCILayer,
+		},
+		{
+			name: "oci image w/ convertable docker layer and zstd compression",
+			fields: fields{
+				image: ociFakeImage{},
+				opts: &config.KanikoOptions{
+					Compression: config.ZStd,
+				},
+			},
+			args: args{
+				layer: fakeLayer{
+					mediaType: types.DockerLayer,
+				},
+			},
+			expectedMediaType: types.OCILayerZStd,
+		},
+		{
+			name: "docker image and oci zstd layer",
+			fields: fields{
+				image: dockerFakeImage{},
+				opts:  &config.KanikoOptions{},
+			},
+			args: args{
+				layer: fakeLayer{
+					mediaType: types.OCILayerZStd,
+				},
+			},
+			expectedMediaType: types.DockerLayer,
+		},
+		{
+			name: "docker image w/ uncovertable oci image",
+			fields: fields{
+				image: dockerFakeImage{},
+				opts:  &config.KanikoOptions{},
+			},
+			args: args{
+				layer: fakeLayer{
+					mediaType: types.OCIUncompressedRestrictedLayer,
+				},
+			},
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			s := &stageBuilder{
+				stage:            tt.fields.stage,
+				image:            tt.fields.image,
+				cf:               tt.fields.cf,
+				baseImageDigest:  tt.fields.baseImageDigest,
+				finalCacheKey:    tt.fields.finalCacheKey,
+				opts:             tt.fields.opts,
+				fileContext:      tt.fields.fileContext,
+				cmds:             tt.fields.cmds,
+				args:             tt.fields.args,
+				crossStageDeps:   tt.fields.crossStageDeps,
+				digestToCacheKey: tt.fields.digestToCacheKey,
+				stageIdxToDigest: tt.fields.stageIdxToDigest,
+				snapshotter:      tt.fields.snapshotter,
+				layerCache:       tt.fields.layerCache,
+				pushLayerToCache: tt.fields.pushLayerToCache,
+			}
+			got, err := s.convertLayerMediaType(tt.args.layer)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("stageBuilder.convertLayerMediaType() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if err == nil {
+				mt, _ := got.MediaType()
+				if mt != tt.expectedMediaType {
+					t.Errorf("stageBuilder.convertLayerMediaType() = %v, want %v", mt, tt.expectedMediaType)
+				}
+			}
+		})
 	}
 }
