@@ -17,9 +17,9 @@ limitations under the License.
 package executor
 
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -41,6 +41,7 @@ import (
 	"github.com/google/go-containerregistry/pkg/v1/remote"
 	"github.com/google/go-containerregistry/pkg/v1/remote/transport"
 	"github.com/google/go-containerregistry/pkg/v1/tarball"
+	"github.com/google/go-containerregistry/pkg/v1/types"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/afero"
@@ -52,11 +53,23 @@ type withUserAgent struct {
 
 // for testing
 var (
-	newRetry = transport.NewRetry
+	newRetry          = transport.NewRetry
+	DummyDestinations = []string{DummyDestination}
 )
 
 const (
 	UpstreamClientUaKey = "UPSTREAM_CLIENT_TYPE"
+	DummyDestination    = "docker.io/unset-repo/unset-image-name"
+)
+
+var (
+	// known tag immutability errors
+	errTagImmutable = []string{
+		// https://cloud.google.com/artifact-registry/docs/docker/troubleshoot#push
+		"The repository has enabled tag immutability",
+		// https://docs.aws.amazon.com/AmazonECR/latest/userguide/image-tag-mutability.html
+		"cannot be overwritten because the repository is immutable",
+	}
 )
 
 func (w *withUserAgent) RoundTrip(r *http.Request) (*http.Response, error) {
@@ -70,7 +83,7 @@ func (w *withUserAgent) RoundTrip(r *http.Request) (*http.Response, error) {
 
 // for testing
 var (
-	fs                        = afero.NewOsFs()
+	newOsFs                   = afero.NewOsFs()
 	checkRemotePushPermission = remote.CheckPushPermission
 )
 
@@ -78,10 +91,19 @@ var (
 // push to every specified destination.
 func CheckPushPermissions(opts *config.KanikoOptions) error {
 	targets := opts.Destinations
-	// When no push is set, whe want to check permissions for the cache repo
-	// instead of the destinations
-	if opts.NoPush {
-		targets = []string{opts.CacheRepo}
+	// When no push and no push cache are set, we don't need to check permissions
+	if opts.SkipPushPermissionCheck {
+		targets = []string{}
+	} else if opts.NoPush && opts.NoPushCache {
+		targets = []string{}
+	} else if opts.NoPush && !opts.NoPushCache {
+		// When no push is set, we want to check permissions for the cache repo
+		// instead of the destinations
+		if isOCILayout(opts.CacheRepo) {
+			targets = []string{} // no need to check push permissions if we're just writing to disk
+		} else {
+			targets = []string{opts.CacheRepo}
+		}
 	}
 
 	checked := map[string]bool{}
@@ -102,7 +124,11 @@ func CheckPushPermissions(opts *config.KanikoOptions) error {
 			}
 			destRef.Repository.Registry = newReg
 		}
-		tr := newRetry(util.MakeTransport(opts.RegistryOptions, registryName))
+		rt, err := util.MakeTransport(opts.RegistryOptions, registryName)
+		if err != nil {
+			return errors.Wrapf(err, "making transport for registry %q", registryName)
+		}
+		tr := newRetry(rt)
 		if err := checkRemotePushPermission(destRef, creds.GetKeychain(), tr); err != nil {
 			return errors.Wrapf(err, "checking push permission for %q", destRef)
 		}
@@ -120,22 +146,40 @@ func getDigest(image v1.Image) ([]byte, error) {
 }
 
 func writeDigestFile(path string, digestByteArray []byte) error {
+	if strings.HasPrefix(path, "https://") {
+		// Do a HTTP PUT to the URL; this could be a pre-signed URL to S3 or GCS or Azure
+		req, err := http.NewRequest("PUT", path, bytes.NewReader(digestByteArray)) //nolint:noctx
+		if err != nil {
+			return err
+		}
+		req.Header.Set("Content-Type", "text/plain")
+		_, err = http.DefaultClient.Do(req)
+		return err
+	}
+
 	parentDir := filepath.Dir(path)
 	if _, err := os.Stat(parentDir); os.IsNotExist(err) {
 		if err := os.MkdirAll(parentDir, 0700); err != nil {
-			logrus.Debugf("error creating %s, %s", parentDir, err)
+			logrus.Debugf("Error creating %s, %s", parentDir, err)
 			return err
 		}
 		logrus.Tracef("Created directory %v", parentDir)
 	}
-	return ioutil.WriteFile(path, digestByteArray, 0644)
+	return os.WriteFile(path, digestByteArray, 0644)
 }
 
-// DoPush is responsible for pushing image to the destinations specified in opts
+// DoPush is responsible for pushing image to the destinations specified in opts.
+// A dummy destination would be set when --no-push is set to true and --tar-path
+// is not empty with empty --destinations.
 func DoPush(image v1.Image, opts *config.KanikoOptions) error {
 	t := timing.Start("Total Push Time")
 	var digestByteArray []byte
 	var builder strings.Builder
+
+	if !opts.NoPush && len(opts.Destinations) == 0 {
+		return errors.New("must provide at least one destination to push")
+	}
+
 	if opts.DigestFile != "" || opts.ImageNameDigestFile != "" || opts.ImageNameTagDigestFile != "" {
 		var err error
 		digestByteArray, err = getDigest(image)
@@ -158,6 +202,12 @@ func DoPush(image v1.Image, opts *config.KanikoOptions) error {
 		}
 		if err := path.AppendImage(image); err != nil {
 			return errors.Wrap(err, "appending image")
+		}
+	}
+
+	if opts.NoPush && len(opts.Destinations) == 0 {
+		if opts.TarPath != "" {
+			setDummyDestinations(opts)
 		}
 	}
 
@@ -196,10 +246,6 @@ func DoPush(image v1.Image, opts *config.KanikoOptions) error {
 	if opts.TarPath != "" {
 		tagToImage := map[name.Tag]v1.Image{}
 
-		if len(destRefs) == 0 {
-			return errors.New("must provide at least one destination when tarPath is specified")
-		}
-
 		for _, destRef := range destRefs {
 			tagToImage[destRef] = image
 		}
@@ -230,7 +276,11 @@ func DoPush(image v1.Image, opts *config.KanikoOptions) error {
 			return errors.Wrap(err, "resolving pushAuth")
 		}
 
-		tr := newRetry(util.MakeTransport(opts.RegistryOptions, registryName))
+		localRt, err := util.MakeTransport(opts.RegistryOptions, registryName)
+		if err != nil {
+			return errors.Wrapf(err, "making transport for registry %q", registryName)
+		}
+		tr := newRetry(localRt)
 		rt := &withUserAgent{t: tr}
 
 		logrus.Infof("Pushing image to %s", destRef.String())
@@ -240,10 +290,23 @@ func DoPush(image v1.Image, opts *config.KanikoOptions) error {
 			if err != nil {
 				return err
 			}
+			digest := destRef.Context().Digest(dig.String())
 			if err := remote.Write(destRef, image, remote.WithAuth(pushAuth), remote.WithTransport(rt)); err != nil {
+				if !opts.PushIgnoreImmutableTagErrors {
+					return err
+				}
+
+				// check for known "tag immutable" errors
+				errStr := err.Error()
+				for _, candidate := range errTagImmutable {
+					if strings.Contains(errStr, candidate) {
+						logrus.Infof("Immutable tag error ignored for %s", digest)
+						return nil
+					}
+				}
 				return err
 			}
-			logrus.Infof("Pushed %s", destRef.Context().Digest(dig.String()))
+			logrus.Infof("Pushed %s", digest)
 			return nil
 		}
 
@@ -260,7 +323,7 @@ func writeImageOutputs(image v1.Image, destRefs []name.Tag) error {
 	if dir == "" {
 		return nil
 	}
-	f, err := fs.Create(filepath.Join(dir, "images"))
+	f, err := newOsFs.Create(filepath.Join(dir, "images"))
 	if err != nil {
 		return err
 	}
@@ -286,19 +349,31 @@ func writeImageOutputs(image v1.Image, destRefs []name.Tag) error {
 	return nil
 }
 
-// pushLayerToCache pushes layer (tagged with cacheKey) to opts.Cache
-// if opts.Cache doesn't exist, infer the cache from the given destination
+// pushLayerToCache pushes layer (tagged with cacheKey) to opts.CacheRepo
+// if opts.CacheRepo doesn't exist, infer the cache from the given destination
 func pushLayerToCache(opts *config.KanikoOptions, cacheKey string, tarPath string, createdBy string) error {
-	var layer v1.Layer
-	var err error
+	var layerOpts []tarball.LayerOption
 	if opts.CompressedCaching == true {
-		layer, err = tarball.LayerFromFile(tarPath, tarball.WithCompressedCaching)
-	} else {
-		layer, err = tarball.LayerFromFile(tarPath)
+		layerOpts = append(layerOpts, tarball.WithCompressedCaching)
 	}
+
+	if opts.CompressionLevel > 0 {
+		layerOpts = append(layerOpts, tarball.WithCompressionLevel(opts.CompressionLevel))
+	}
+
+	switch opts.Compression {
+	case config.ZStd:
+		layerOpts = append(layerOpts, tarball.WithCompression("zstd"), tarball.WithMediaType(types.OCILayerZStd))
+
+	case config.GZip:
+		// layer already gzipped by default
+	}
+
+	layer, err := tarball.LayerFromFile(tarPath, layerOpts...)
 	if err != nil {
 		return err
 	}
+
 	cache, err := cache.Destination(opts, cacheKey)
 	if err != nil {
 		return errors.Wrap(err, "getting cache destination")
@@ -323,10 +398,20 @@ func pushLayerToCache(opts *config.KanikoOptions, cacheKey string, tarPath strin
 		return errors.Wrap(err, "appending layer onto empty image")
 	}
 	cacheOpts := *opts
-	cacheOpts.TarPath = ""   // tarPath doesn't make sense for Docker layers
-	cacheOpts.NoPush = false // we want to push cached layers
+	cacheOpts.TarPath = ""              // tarPath doesn't make sense for Docker layers
+	cacheOpts.NoPush = opts.NoPushCache // we do not want to push cache if --no-push-cache is set.
 	cacheOpts.Destinations = []string{cache}
 	cacheOpts.InsecureRegistries = opts.InsecureRegistries
 	cacheOpts.SkipTLSVerifyRegistries = opts.SkipTLSVerifyRegistries
+	if isOCILayout(cache) {
+		cacheOpts.OCILayoutPath = strings.TrimPrefix(cache, "oci:")
+		cacheOpts.NoPush = true
+	}
 	return DoPush(empty, &cacheOpts)
+}
+
+// setDummyDestinations sets the dummy destinations required to generate new
+// tag names for tarPath in DoPush.
+func setDummyDestinations(opts *config.KanikoOptions) {
+	opts.Destinations = DummyDestinations
 }
