@@ -18,7 +18,6 @@ package filesystem
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"reflect"
@@ -51,12 +50,7 @@ func Test_ResolvePaths(t *testing.T) {
 	}
 
 	t.Run("list of files", func(t *testing.T) {
-		dir, err := ioutil.TempDir("", "snapshot-test")
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		defer os.RemoveAll(dir)
+		dir := t.TempDir()
 
 		files := []string{
 			"/foo/bar.txt",
@@ -72,7 +66,7 @@ func Test_ResolvePaths(t *testing.T) {
 					t.Fatal(err)
 				}
 
-				if err := ioutil.WriteFile(fTarget, []byte{}, 0777); err != nil {
+				if err := os.WriteFile(fTarget, []byte{}, 0777); err != nil {
 					t.Fatal(err)
 				}
 
@@ -148,7 +142,7 @@ func Test_ResolvePaths(t *testing.T) {
 					t.Fatal(err)
 				}
 
-				if err := ioutil.WriteFile(filepath.Join(target, "meow.txt"), []byte{}, 0777); err != nil {
+				if err := os.WriteFile(filepath.Join(target, "meow.txt"), []byte{}, 0777); err != nil {
 					t.Fatal(err)
 				}
 
@@ -187,10 +181,7 @@ func Test_ResolvePaths(t *testing.T) {
 
 func Test_resolveSymlinkAncestor(t *testing.T) {
 	setupDirs := func(t *testing.T) (string, string) {
-		testDir, err := ioutil.TempDir("", "")
-		if err != nil {
-			t.Fatal(err)
-		}
+		testDir := t.TempDir()
 
 		targetDir := filepath.Join(testDir, "bar", "baz")
 
@@ -200,7 +191,7 @@ func Test_resolveSymlinkAncestor(t *testing.T) {
 
 		targetPath := filepath.Join(targetDir, "bam.txt")
 
-		if err := ioutil.WriteFile(targetPath, []byte("meow"), 0777); err != nil {
+		if err := os.WriteFile(targetPath, []byte("meow"), 0777); err != nil {
 			t.Fatal(err)
 		}
 
