@@ -13,12 +13,12 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package snapshot
 
 import (
 	"archive/tar"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"sort"
@@ -32,7 +32,7 @@ import (
 )
 
 func TestSnapshotFSFileChange(t *testing.T) {
-	testDir, snapshotter, cleanup, err := setUpTest()
+	testDir, snapshotter, cleanup, err := setUpTest(t)
 	testDirWithoutLeadingSlash := strings.TrimLeft(testDir, "/")
 	defer cleanup()
 	if err != nil {
@@ -75,7 +75,7 @@ func TestSnapshotFSFileChange(t *testing.T) {
 	actualFiles := []string{}
 	for {
 		hdr, err := tr.Next()
-		if err == io.EOF {
+		if errors.Is(err, io.EOF) {
 			break
 		}
 
@@ -87,7 +87,7 @@ func TestSnapshotFSFileChange(t *testing.T) {
 		if hdr.Typeflag == tar.TypeDir {
 			continue
 		}
-		contents, _ := ioutil.ReadAll(tr)
+		contents, _ := io.ReadAll(tr)
 		if string(contents) != snapshotFiles[hdr.Name] {
 			t.Fatalf("Contents of %s incorrect, expected: %s, actual: %s", hdr.Name, snapshotFiles[hdr.Name], string(contents))
 		}
@@ -98,7 +98,7 @@ func TestSnapshotFSFileChange(t *testing.T) {
 }
 
 func TestSnapshotFSIsReproducible(t *testing.T) {
-	testDir, snapshotter, cleanup, err := setUpTest()
+	testDir, snapshotter, cleanup, err := setUpTest(t)
 	defer cleanup()
 	if err != nil {
 		t.Fatal(err)
@@ -117,19 +117,10 @@ func TestSnapshotFSIsReproducible(t *testing.T) {
 		t.Fatalf("Error taking snapshot of fs: %s", err)
 	}
 
-	f, err := os.Open(tarPath)
+	// Check contents of the snapshot, make sure contents are sorted by name
+	filesInTar, err := listFilesInTar(tarPath)
 	if err != nil {
 		t.Fatal(err)
-	}
-	// Check contents of the snapshot, make sure contents are sorted by name
-	tr := tar.NewReader(f)
-	var filesInTar []string
-	for {
-		hdr, err := tr.Next()
-		if err == io.EOF {
-			break
-		}
-		filesInTar = append(filesInTar, hdr.Name)
 	}
 	if !sort.StringsAreSorted(filesInTar) {
 		t.Fatalf("Expected the file in the tar archive were sorted, actual list was not sorted: %v", filesInTar)
@@ -137,7 +128,7 @@ func TestSnapshotFSIsReproducible(t *testing.T) {
 }
 
 func TestSnapshotFSChangePermissions(t *testing.T) {
-	testDir, snapshotter, cleanup, err := setUpTest()
+	testDir, snapshotter, cleanup, err := setUpTest(t)
 	testDirWithoutLeadingSlash := strings.TrimLeft(testDir, "/")
 	defer cleanup()
 	if err != nil {
@@ -174,7 +165,7 @@ func TestSnapshotFSChangePermissions(t *testing.T) {
 	foundFiles := []string{}
 	for {
 		hdr, err := tr.Next()
-		if err == io.EOF {
+		if errors.Is(err, io.EOF) {
 			break
 		}
 		foundFiles = append(foundFiles, hdr.Name)
@@ -184,7 +175,7 @@ func TestSnapshotFSChangePermissions(t *testing.T) {
 		if hdr.Typeflag == tar.TypeDir {
 			continue
 		}
-		contents, _ := ioutil.ReadAll(tr)
+		contents, _ := io.ReadAll(tr)
 		if string(contents) != snapshotFiles[hdr.Name] {
 			t.Fatalf("Contents of %s incorrect, expected: %s, actual: %s", hdr.Name, snapshotFiles[hdr.Name], string(contents))
 		}
@@ -197,8 +188,53 @@ func TestSnapshotFSChangePermissions(t *testing.T) {
 	}
 }
 
+func TestSnapshotFSReplaceDirWithLink(t *testing.T) {
+	testDir, snapshotter, cleanup, err := setUpTest(t)
+	defer cleanup()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// replace non-empty directory "bar" with link to file "foo"
+	bar := filepath.Join(testDir, "bar")
+	err = os.RemoveAll(bar)
+	if err != nil {
+		t.Fatal(err)
+	}
+	foo := filepath.Join(testDir, "foo")
+	err = os.Symlink(foo, bar)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	tarPath, err := snapshotter.TakeSnapshotFS()
+	if err != nil {
+		t.Fatalf("Error taking snapshot of fs: %s", err)
+	}
+
+	actualFiles, err := listFilesInTar(tarPath)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Expect "bar", which used to be a non-empty directory but now is a symlink. We don't want whiteout files for
+	// the deleted files in bar, because without a parent directory for them the tar cannot be extracted.
+	testDirWithoutLeadingSlash := strings.TrimLeft(testDir, "/")
+	expectedFiles := []string{
+		filepath.Join(testDirWithoutLeadingSlash, "bar"),
+		filepath.Join(testDirWithoutLeadingSlash, "foo"),
+	}
+	for _, path := range util.ParentDirectoriesWithoutLeadingSlash(filepath.Join(testDir, "foo")) {
+		expectedFiles = append(expectedFiles, strings.TrimRight(path, "/")+"/")
+	}
+
+	sort.Strings(expectedFiles)
+	sort.Strings(actualFiles)
+	testutil.CheckErrorAndDeepEqual(t, false, nil, expectedFiles, actualFiles)
+}
+
 func TestSnapshotFiles(t *testing.T) {
-	testDir, snapshotter, cleanup, err := setUpTest()
+	testDir, snapshotter, cleanup, err := setUpTest(t)
 	testDirWithoutLeadingSlash := strings.TrimLeft(testDir, "/")
 	defer cleanup()
 	if err != nil {
@@ -227,30 +263,19 @@ func TestSnapshotFiles(t *testing.T) {
 		expectedFiles = append(expectedFiles, strings.TrimRight(path, "/")+"/")
 	}
 
-	f, err := os.Open(tarPath)
+	// Check contents of the snapshot, make sure contents is equivalent to snapshotFiles
+	actualFiles, err := listFilesInTar(tarPath)
 	if err != nil {
 		t.Fatal(err)
 	}
-	// Check contents of the snapshot, make sure contents is equivalent to snapshotFiles
-	tr := tar.NewReader(f)
-	var actualFiles []string
-	for {
-		hdr, err := tr.Next()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			t.Fatal(err)
-		}
-		actualFiles = append(actualFiles, hdr.Name)
-	}
+
 	sort.Strings(expectedFiles)
 	sort.Strings(actualFiles)
 	testutil.CheckErrorAndDeepEqual(t, false, nil, expectedFiles, actualFiles)
 }
 
 func TestEmptySnapshotFS(t *testing.T) {
-	_, snapshotter, cleanup, err := setUpTest()
+	_, snapshotter, cleanup, err := setUpTest(t)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -268,7 +293,7 @@ func TestEmptySnapshotFS(t *testing.T) {
 	}
 	tr := tar.NewReader(f)
 
-	if _, err := tr.Next(); err != io.EOF {
+	if _, err := tr.Next(); !errors.Is(err, io.EOF) {
 		t.Fatal("no files expected in tar, found files.")
 	}
 }
@@ -305,11 +330,10 @@ func TestFileWithLinks(t *testing.T) {
 
 	for _, tt := range tcs {
 		t.Run(tt.name, func(t *testing.T) {
-			testDir, cleanup, err := setUpTestDir()
+			testDir, err := setUpTestDir(t)
 			if err != nil {
 				t.Fatal(err)
 			}
-			defer cleanup()
 			if err := setupSymlink(testDir, link, tt.linkFileTarget); err != nil {
 				t.Fatalf("could not set up symlink due to %s", err)
 			}
@@ -322,7 +346,7 @@ func TestFileWithLinks(t *testing.T) {
 	}
 }
 
-func TestSnasphotPreservesFileOrder(t *testing.T) {
+func TestSnapshotPreservesFileOrder(t *testing.T) {
 	newFiles := map[string]string{
 		"foo":     "newbaz1",
 		"bar/bat": "baz",
@@ -343,7 +367,7 @@ func TestSnasphotPreservesFileOrder(t *testing.T) {
 	filesInTars := [][]string{}
 
 	for i := 0; i <= 2; i++ {
-		testDir, snapshotter, cleanup, err := setUpTest()
+		testDir, snapshotter, cleanup, err := setUpTest(t)
 		testDirWithoutLeadingSlash := strings.TrimLeft(testDir, "/")
 		defer cleanup()
 
@@ -367,21 +391,14 @@ func TestSnasphotPreservesFileOrder(t *testing.T) {
 			t.Fatalf("Error taking snapshot of fs: %s", err)
 		}
 
-		f, err := os.Open(tarPath)
+		filesInTar, err := listFilesInTar(tarPath)
 		if err != nil {
 			t.Fatal(err)
 		}
-		tr := tar.NewReader(f)
+
 		filesInTars = append(filesInTars, []string{})
-		for {
-			hdr, err := tr.Next()
-			if err == io.EOF {
-				break
-			}
-			if err != nil {
-				t.Fatal(err)
-			}
-			filesInTars[i] = append(filesInTars[i], strings.TrimPrefix(hdr.Name, testDirWithoutLeadingSlash))
+		for _, fn := range filesInTar {
+			filesInTars[i] = append(filesInTars[i], strings.TrimPrefix(fn, testDirWithoutLeadingSlash))
 		}
 	}
 
@@ -392,7 +409,7 @@ func TestSnasphotPreservesFileOrder(t *testing.T) {
 }
 
 func TestSnapshotWithForceBuildMetadataSet(t *testing.T) {
-	_, snapshotter, cleanup, err := setUpTest()
+	_, snapshotter, cleanup, err := setUpTest(t)
 	defer cleanup()
 
 	if err != nil {
@@ -412,7 +429,7 @@ func TestSnapshotWithForceBuildMetadataSet(t *testing.T) {
 }
 
 func TestSnapshotWithForceBuildMetadataIsNotSet(t *testing.T) {
-	_, snapshotter, cleanup, err := setUpTest()
+	_, snapshotter, cleanup, err := setUpTest(t)
 	defer cleanup()
 
 	if err != nil {
@@ -431,7 +448,68 @@ func TestSnapshotWithForceBuildMetadataIsNotSet(t *testing.T) {
 	}
 }
 
-func TestSnasphotPreservesWhiteoutOrder(t *testing.T) {
+func TestSnapshotIncludesParentDirBeforeWhiteoutFile(t *testing.T) {
+	testDir, snapshotter, cleanup, err := setUpTest(t)
+	defer cleanup()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Take a snapshot
+	filesToSnapshot := []string{filepath.Join(testDir, "kaniko/file", "bar/bat")}
+	_, err = snapshotter.TakeSnapshot(filesToSnapshot, false, false)
+	if err != nil {
+		t.Fatalf("Error taking snapshot of fs: %s", err)
+	}
+
+	// Add a file
+	newFiles := map[string]string{
+		"kaniko/new-file": "baz",
+	}
+	if err := testutil.SetupFiles(testDir, newFiles); err != nil {
+		t.Fatalf("Error setting up fs: %s", err)
+	}
+	filesToSnapshot = append(filesToSnapshot, filepath.Join(testDir, "kaniko/new-file"))
+
+	// Delete files
+	filesToDelete := []string{"kaniko/file", "bar"}
+	for _, fn := range filesToDelete {
+		err = os.RemoveAll(filepath.Join(testDir, fn))
+		if err != nil {
+			t.Fatalf("Error deleting file: %s", err)
+		}
+	}
+
+	// Take a snapshot again
+	tarPath, err := snapshotter.TakeSnapshot(filesToSnapshot, true, false)
+	if err != nil {
+		t.Fatalf("Error taking snapshot of fs: %s", err)
+	}
+
+	actualFiles, err := listFilesInTar(tarPath)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	testDirWithoutLeadingSlash := strings.TrimLeft(testDir, "/")
+	expectedFiles := []string{
+		filepath.Join(testDirWithoutLeadingSlash, "kaniko/.wh.file"),
+		filepath.Join(testDirWithoutLeadingSlash, "kaniko/new-file"),
+		filepath.Join(testDirWithoutLeadingSlash, ".wh.bar"),
+		"/",
+	}
+	for parentDir := filepath.Dir(expectedFiles[0]); parentDir != "."; parentDir = filepath.Dir(parentDir) {
+		expectedFiles = append(expectedFiles, parentDir+"/")
+	}
+
+	// Sorting does the right thing in this case. The expected order for a directory is:
+	// Parent dirs first, then whiteout files in the directory, then other files in that directory
+	sort.Strings(expectedFiles)
+
+	testutil.CheckErrorAndDeepEqual(t, false, nil, expectedFiles, actualFiles)
+}
+
+func TestSnapshotPreservesWhiteoutOrder(t *testing.T) {
 	newFiles := map[string]string{
 		"foo":     "newbaz1",
 		"bar/bat": "baz",
@@ -452,7 +530,7 @@ func TestSnasphotPreservesWhiteoutOrder(t *testing.T) {
 	filesInTars := [][]string{}
 
 	for i := 0; i <= 2; i++ {
-		testDir, snapshotter, cleanup, err := setUpTest()
+		testDir, snapshotter, cleanup, err := setUpTest(t)
 		testDirWithoutLeadingSlash := strings.TrimLeft(testDir, "/")
 		defer cleanup()
 
@@ -489,21 +567,14 @@ func TestSnasphotPreservesWhiteoutOrder(t *testing.T) {
 			t.Fatalf("Error taking snapshot of fs: %s", err)
 		}
 
-		f, err := os.Open(tarPath)
+		filesInTar, err := listFilesInTar(tarPath)
 		if err != nil {
 			t.Fatal(err)
 		}
-		tr := tar.NewReader(f)
+
 		filesInTars = append(filesInTars, []string{})
-		for {
-			hdr, err := tr.Next()
-			if err == io.EOF {
-				break
-			}
-			if err != nil {
-				t.Fatal(err)
-			}
-			filesInTars[i] = append(filesInTars[i], strings.TrimPrefix(hdr.Name, testDirWithoutLeadingSlash))
+		for _, fn := range filesInTar {
+			filesInTars[i] = append(filesInTars[i], strings.TrimPrefix(fn, testDirWithoutLeadingSlash))
 		}
 	}
 
@@ -514,7 +585,7 @@ func TestSnasphotPreservesWhiteoutOrder(t *testing.T) {
 }
 
 func TestSnapshotOmitsUnameGname(t *testing.T) {
-	_, snapshotter, cleanup, err := setUpTest()
+	_, snapshotter, cleanup, err := setUpTest(t)
 
 	defer cleanup()
 	if err != nil {
@@ -533,7 +604,7 @@ func TestSnapshotOmitsUnameGname(t *testing.T) {
 	tr := tar.NewReader(f)
 	for {
 		hdr, err := tr.Next()
-		if err == io.EOF {
+		if errors.Is(err, io.EOF) {
 			break
 		}
 		if err != nil {
@@ -560,11 +631,8 @@ func sortAndCompareFilepaths(t *testing.T, testDir string, expected []string, ac
 	testutil.CheckDeepEqual(t, expectedFullPaths, actual)
 }
 
-func setUpTestDir() (string, func(), error) {
-	testDir, err := ioutil.TempDir("", "")
-	if err != nil {
-		return "", nil, errors.Wrap(err, "setting up temp dir")
-	}
+func setUpTestDir(t *testing.T) (string, error) {
+	testDir := t.TempDir()
 	files := map[string]string{
 		"foo":         "baz1",
 		"bar/bat":     "baz2",
@@ -573,30 +641,23 @@ func setUpTestDir() (string, func(), error) {
 	}
 	// Set up initial files
 	if err := testutil.SetupFiles(testDir, files); err != nil {
-		return "", nil, errors.Wrap(err, "setting up file system")
+		return "", errors.Wrap(err, "setting up file system")
 	}
 
-	cleanup := func() {
-		os.RemoveAll(testDir)
-	}
-
-	return testDir, cleanup, nil
+	return testDir, nil
 }
 
-func setUpTest() (string, *Snapshotter, func(), error) {
-	testDir, dirCleanUp, err := setUpTestDir()
+func setUpTest(t *testing.T) (string, *Snapshotter, func(), error) {
+	testDir, err := setUpTestDir(t)
 	if err != nil {
 		return "", nil, nil, err
 	}
-	snapshotPath, err := ioutil.TempDir("", "")
-	if err != nil {
-		return "", nil, nil, errors.Wrap(err, "setting up temp dir")
-	}
 
+	snapshotPath := t.TempDir()
 	snapshotPathPrefix = snapshotPath
 
 	// Take the initial snapshot
-	l := NewLayeredMap(util.Hasher(), util.CacheHasher())
+	l := NewLayeredMap(util.Hasher())
 	snapshotter := NewSnapshotter(l, testDir)
 	if err := snapshotter.Init(); err != nil {
 		return "", nil, nil, errors.Wrap(err, "initializing snapshotter")
@@ -605,10 +666,28 @@ func setUpTest() (string, *Snapshotter, func(), error) {
 	original := config.KanikoDir
 	config.KanikoDir = testDir
 	cleanup := func() {
-		os.RemoveAll(snapshotPath)
 		config.KanikoDir = original
-		dirCleanUp()
 	}
 
 	return testDir, snapshotter, cleanup, nil
+}
+
+func listFilesInTar(path string) ([]string, error) {
+	f, err := os.Open(path)
+	if err != nil {
+		return nil, err
+	}
+	tr := tar.NewReader(f)
+	var files []string
+	for {
+		hdr, err := tr.Next()
+		if errors.Is(err, io.EOF) {
+			break
+		}
+		if err != nil {
+			return nil, err
+		}
+		files = append(files, hdr.Name)
+	}
+	return files, nil
 }
