@@ -21,7 +21,7 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"io/ioutil"
+	"io/fs"
 	"os"
 	"path/filepath"
 	"reflect"
@@ -40,10 +40,7 @@ import (
 )
 
 func Test_DetectFilesystemSkiplist(t *testing.T) {
-	testDir, err := ioutil.TempDir("", "")
-	if err != nil {
-		t.Fatalf("Error creating tempdir: %s", err)
-	}
+	testDir := t.TempDir()
 	fileContents := `
 	228 122 0:90 / / rw,relatime - aufs none rw,si=f8e2406af90782bc,dio,dirperm1
 	229 228 0:98 / /proc rw,nosuid,nodev,noexec,relatime - proc proc rw
@@ -52,14 +49,14 @@ func Test_DetectFilesystemSkiplist(t *testing.T) {
 	232 228 0:101 / /sys ro,nosuid,nodev,noexec,relatime - sysfs sysfs ro`
 
 	path := filepath.Join(testDir, "mountinfo")
-	if err := os.MkdirAll(filepath.Dir(path), 0750); err != nil {
+	if err := os.MkdirAll(filepath.Dir(path), 0o750); err != nil {
 		t.Fatalf("Error creating tempdir: %s", err)
 	}
-	if err := ioutil.WriteFile(path, []byte(fileContents), 0644); err != nil {
+	if err := os.WriteFile(path, []byte(fileContents), 0o644); err != nil {
 		t.Fatalf("Error writing file contents to %s: %s", path, err)
 	}
 
-	err = DetectFilesystemIgnoreList(path)
+	err := DetectFilesystemIgnoreList(path)
 	expectedSkiplist := []IgnoreListEntry{
 		{"/kaniko", false},
 		{"/proc", false},
@@ -155,11 +152,7 @@ var tests = []struct {
 
 func Test_RelativeFiles(t *testing.T) {
 	for _, test := range tests {
-		testDir, err := ioutil.TempDir("", "")
-		if err != nil {
-			t.Fatalf("err setting up temp dir: %v", err)
-		}
-		defer os.RemoveAll(testDir)
+		testDir := t.TempDir()
 		if err := testutil.SetupFiles(testDir, test.files); err != nil {
 			t.Fatalf("err setting up files: %v", err)
 		}
@@ -499,7 +492,7 @@ func fileExists(p string) checker {
 
 func fileMatches(p string, c []byte) checker {
 	return func(root string, t *testing.T) {
-		actual, err := ioutil.ReadFile(filepath.Join(root, p))
+		actual, err := os.ReadFile(filepath.Join(root, p))
 		if err != nil {
 			t.Fatalf("error reading file: %s", p)
 		}
@@ -659,11 +652,7 @@ func Test_UnTar(t *testing.T) {
 	}
 	for _, tc := range tcs {
 		t.Run(tc.name, func(t *testing.T) {
-			testDir, err := ioutil.TempDir("", "")
-			if err != nil {
-				t.Fatal(err)
-			}
-			defer os.RemoveAll(testDir)
+			testDir := t.TempDir()
 			if err := createUncompressedTar(tc.setupTarContents, tc.tarFileName, testDir); err != nil {
 				t.Fatal(err)
 			}
@@ -683,7 +672,6 @@ func Test_UnTar(t *testing.T) {
 			sort.Strings(tc.expectedFileList)
 			sort.Strings(fileList)
 			testutil.CheckErrorAndDeepEqual(t, tc.errorExpected, err, tc.expectedFileList, fileList)
-
 		})
 	}
 }
@@ -706,37 +694,37 @@ func TestExtractFile(t *testing.T) {
 		{
 			name:     "normal file",
 			contents: []byte("helloworld"),
-			hdrs:     []*tar.Header{fileHeader("./bar", "helloworld", 0644, defaultTestTime)},
+			hdrs:     []*tar.Header{fileHeader("./bar", "helloworld", 0o644, defaultTestTime)},
 			checkers: []checker{
 				fileExists("/bar"),
 				fileMatches("/bar", []byte("helloworld")),
-				permissionsMatch("/bar", 0644),
+				permissionsMatch("/bar", 0o644),
 				timesMatch("/bar", defaultTestTime),
 			},
 		},
 		{
 			name:     "normal file, directory does not exist",
 			contents: []byte("helloworld"),
-			hdrs:     []*tar.Header{fileHeader("./foo/bar", "helloworld", 0644, defaultTestTime)},
+			hdrs:     []*tar.Header{fileHeader("./foo/bar", "helloworld", 0o644, defaultTestTime)},
 			checkers: []checker{
 				fileExists("/foo/bar"),
 				fileMatches("/foo/bar", []byte("helloworld")),
-				permissionsMatch("/foo/bar", 0644),
-				permissionsMatch("/foo", 0755|os.ModeDir),
+				permissionsMatch("/foo/bar", 0o644),
+				permissionsMatch("/foo", 0o755|os.ModeDir),
 			},
 		},
 		{
 			name:     "normal file, directory is created after",
 			contents: []byte("helloworld"),
 			hdrs: []*tar.Header{
-				fileHeader("./foo/bar", "helloworld", 0644, defaultTestTime),
-				dirHeader("./foo", 0722),
+				fileHeader("./foo/bar", "helloworld", 0o644, defaultTestTime),
+				dirHeader("./foo", 0o722),
 			},
 			checkers: []checker{
 				fileExists("/foo/bar"),
 				fileMatches("/foo/bar", []byte("helloworld")),
-				permissionsMatch("/foo/bar", 0644),
-				permissionsMatch("/foo", 0722|os.ModeDir),
+				permissionsMatch("/foo/bar", 0o644),
+				permissionsMatch("/foo", 0o722|os.ModeDir),
 			},
 		},
 		{
@@ -765,15 +753,15 @@ func TestExtractFile(t *testing.T) {
 			hdrs: []*tar.Header{linkHeader("./foo/bar/baz", "../../bat")},
 			checkers: []checker{
 				linkPointsTo("/foo/bar/baz", "../../bat"),
-				permissionsMatch("/foo", 0755|os.ModeDir),
-				permissionsMatch("/foo/bar", 0755|os.ModeDir),
+				permissionsMatch("/foo", 0o755|os.ModeDir),
+				permissionsMatch("/foo/bar", 0o755|os.ModeDir),
 			},
 		},
 		{
 			name:   "hardlink",
 			tmpdir: "/tmp/hardlink",
 			hdrs: []*tar.Header{
-				fileHeader("/bin/gzip", "gzip-binary", 0751, defaultTestTime),
+				fileHeader("/bin/gzip", "gzip-binary", 0o751, defaultTestTime),
 				hardlinkHeader("/bin/uncompress", "/bin/gzip"),
 			},
 			checkers: []checker{
@@ -784,25 +772,25 @@ func TestExtractFile(t *testing.T) {
 		{
 			name:     "file with setuid bit",
 			contents: []byte("helloworld"),
-			hdrs:     []*tar.Header{fileHeader("./bar", "helloworld", 04644, defaultTestTime)},
+			hdrs:     []*tar.Header{fileHeader("./bar", "helloworld", 0o4644, defaultTestTime)},
 			checkers: []checker{
 				fileExists("/bar"),
 				fileMatches("/bar", []byte("helloworld")),
-				permissionsMatch("/bar", 0644|os.ModeSetuid),
+				permissionsMatch("/bar", 0o644|os.ModeSetuid),
 			},
 		},
 		{
 			name:     "dir with sticky bit",
 			contents: []byte("helloworld"),
 			hdrs: []*tar.Header{
-				dirHeader("./foo", 01755),
-				fileHeader("./foo/bar", "helloworld", 0644, defaultTestTime),
+				dirHeader("./foo", 0o1755),
+				fileHeader("./foo/bar", "helloworld", 0o644, defaultTestTime),
 			},
 			checkers: []checker{
 				fileExists("/foo/bar"),
 				fileMatches("/foo/bar", []byte("helloworld")),
-				permissionsMatch("/foo/bar", 0644),
-				permissionsMatch("/foo", 0755|os.ModeDir|os.ModeSticky),
+				permissionsMatch("/foo/bar", 0o644),
+				permissionsMatch("/foo", 0o755|os.ModeDir|os.ModeSticky),
 			},
 		},
 	}
@@ -812,20 +800,16 @@ func TestExtractFile(t *testing.T) {
 			tc := tc
 			t.Parallel()
 			r := ""
-			var err error
 
 			if tc.tmpdir != "" {
 				r = tc.tmpdir
 			} else {
-				r, err = ioutil.TempDir("", "")
-				if err != nil {
-					t.Fatal(err)
-				}
+				r = t.TempDir()
 			}
 			defer os.RemoveAll(r)
 
 			for _, hdr := range tc.hdrs {
-				if err := ExtractFile(r, hdr, bytes.NewReader(tc.contents)); err != nil {
+				if err := ExtractFile(r, hdr, filepath.Clean(hdr.Name), bytes.NewReader(tc.contents)); err != nil {
 					t.Fatal(err)
 				}
 			}
@@ -855,7 +839,7 @@ func TestCopySymlink(t *testing.T) {
 		linkTarget: "/abs/dest",
 		dest:       "overwrite_me",
 		beforeLink: func(r string) error {
-			return ioutil.WriteFile(filepath.Join(r, "overwrite_me"), nil, 0644)
+			return os.WriteFile(filepath.Join(r, "overwrite_me"), nil, 0o644)
 		},
 	}}
 
@@ -863,14 +847,10 @@ func TestCopySymlink(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			tc := tc
 			t.Parallel()
-			r, err := ioutil.TempDir("", "")
-			os.MkdirAll(filepath.Join(r, filepath.Dir(tc.linkTarget)), 0777)
+			r := t.TempDir()
+			os.MkdirAll(filepath.Join(r, filepath.Dir(tc.linkTarget)), 0o777)
 			tc.linkTarget = filepath.Join(r, tc.linkTarget)
-			ioutil.WriteFile(tc.linkTarget, nil, 0644)
-			if err != nil {
-				t.Fatal(err)
-			}
-			defer os.RemoveAll(r)
+			os.WriteFile(tc.linkTarget, nil, 0o644)
 
 			if tc.beforeLink != nil {
 				if err := tc.beforeLink(r); err != nil {
@@ -994,23 +974,20 @@ func Test_correctDockerignoreFileIsUsed(t *testing.T) {
 
 func Test_CopyFile_skips_self(t *testing.T) {
 	t.Parallel()
-	tempDir, err := ioutil.TempDir("", "kaniko_test")
-	if err != nil {
-		t.Fatal(err)
-	}
+	tempDir := t.TempDir()
 
 	tempFile := filepath.Join(tempDir, "foo")
 	expected := "bar"
 
-	if err := ioutil.WriteFile(
+	if err := os.WriteFile(
 		tempFile,
 		[]byte(expected),
-		0755,
+		0o755,
 	); err != nil {
 		t.Fatal(err)
 	}
 
-	ignored, err := CopyFile(tempFile, tempFile, FileContext{}, DoNotChangeUID, DoNotChangeGID)
+	ignored, err := CopyFile(tempFile, tempFile, FileContext{}, DoNotChangeUID, DoNotChangeGID, fs.FileMode(0o600), true)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -1020,7 +997,7 @@ func Test_CopyFile_skips_self(t *testing.T) {
 	}
 
 	// Ensure file has expected contents
-	actualData, err := ioutil.ReadFile(tempFile)
+	actualData, err := os.ReadFile(tempFile)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -1030,24 +1007,22 @@ func Test_CopyFile_skips_self(t *testing.T) {
 	}
 }
 
-func fakeExtract(dest string, hdr *tar.Header, tr io.Reader) error {
+func fakeExtract(_ string, _ *tar.Header, _ string, _ io.Reader) error {
 	return nil
 }
 
 func Test_GetFSFromLayers_with_whiteouts_include_whiteout_enabled(t *testing.T) {
+	resetMountInfoFile := provideEmptyMountinfoFile()
+	defer resetMountInfoFile()
+
 	ctrl := gomock.NewController(t)
 
-	root, err := ioutil.TempDir("", "layers-test")
-	if err != nil {
-		t.Fatal(err)
-	}
+	root := t.TempDir()
 	// Write a whiteout path
 	d1 := []byte("Hello World\n")
-	if err := ioutil.WriteFile(filepath.Join(root, "foobar"), d1, 0644); err != nil {
+	if err := os.WriteFile(filepath.Join(root, "foobar"), d1, 0o644); err != nil {
 		t.Fatal(err)
 	}
-
-	defer os.Remove(root)
 
 	opts := []FSOpt{
 		// I'd rather use the real func (util.ExtractFile)
@@ -1064,7 +1039,7 @@ func Test_GetFSFromLayers_with_whiteouts_include_whiteout_enabled(t *testing.T) 
 
 			hdr := &tar.Header{
 				Name: f,
-				Mode: 0644,
+				Mode: 0o644,
 				Size: int64(len("Hello World\n")),
 			}
 
@@ -1094,7 +1069,7 @@ func Test_GetFSFromLayers_with_whiteouts_include_whiteout_enabled(t *testing.T) 
 	mockLayer := mockv1.NewMockLayer(ctrl)
 	mockLayer.EXPECT().MediaType().Return(types.OCILayer, nil)
 
-	rc := ioutil.NopCloser(buf)
+	rc := io.NopCloser(buf)
 	mockLayer.EXPECT().Uncompressed().Return(rc, nil)
 
 	secondLayerFiles := []string{
@@ -1109,7 +1084,7 @@ func Test_GetFSFromLayers_with_whiteouts_include_whiteout_enabled(t *testing.T) 
 	mockLayer2 := mockv1.NewMockLayer(ctrl)
 	mockLayer2.EXPECT().MediaType().Return(types.OCILayer, nil)
 
-	rc = ioutil.NopCloser(buf)
+	rc = io.NopCloser(buf)
 	mockLayer2.EXPECT().Uncompressed().Return(rc, nil)
 
 	layers := []v1.Layer{
@@ -1135,19 +1110,28 @@ func Test_GetFSFromLayers_with_whiteouts_include_whiteout_enabled(t *testing.T) 
 	}
 }
 
+func provideEmptyMountinfoFile() func() {
+	// Provide empty mountinfo file to prevent /tmp from ending up in ignore list on
+	// distributions with /tmp mountpoint. Otherwise, tests expecting operations in /tmp
+	// can fail.
+	config.MountInfoPath = "/dev/null"
+	return func() {
+		config.MountInfoPath = constants.MountInfoPath
+	}
+}
+
 func Test_GetFSFromLayers_with_whiteouts_include_whiteout_disabled(t *testing.T) {
+	resetMountInfoFile := provideEmptyMountinfoFile()
+	defer resetMountInfoFile()
+
 	ctrl := gomock.NewController(t)
 
-	root, err := ioutil.TempDir("", "layers-test")
-	if err != nil {
-		t.Fatal(err)
-	}
+	root := t.TempDir()
 	// Write a whiteout path
 	d1 := []byte("Hello World\n")
-	if err := ioutil.WriteFile(filepath.Join(root, "foobar"), d1, 0644); err != nil {
+	if err := os.WriteFile(filepath.Join(root, "foobar"), d1, 0o644); err != nil {
 		t.Fatal(err)
 	}
-	defer os.Remove(root)
 
 	opts := []FSOpt{
 		// I'd rather use the real func (util.ExtractFile)
@@ -1163,7 +1147,7 @@ func Test_GetFSFromLayers_with_whiteouts_include_whiteout_disabled(t *testing.T)
 
 			hdr := &tar.Header{
 				Name: f,
-				Mode: 0644,
+				Mode: 0o644,
 				Size: int64(len("Hello world\n")),
 			}
 
@@ -1200,7 +1184,7 @@ func Test_GetFSFromLayers_with_whiteouts_include_whiteout_disabled(t *testing.T)
 
 	f(layerFiles, tw)
 
-	rc := ioutil.NopCloser(buf)
+	rc := io.NopCloser(buf)
 	mockLayer.EXPECT().Uncompressed().Return(rc, nil)
 
 	secondLayerFiles := []string{
@@ -1215,7 +1199,7 @@ func Test_GetFSFromLayers_with_whiteouts_include_whiteout_disabled(t *testing.T)
 	mockLayer2 := mockv1.NewMockLayer(ctrl)
 	mockLayer2.EXPECT().MediaType().Return(types.OCILayer, nil)
 
-	rc = ioutil.NopCloser(buf)
+	rc = io.NopCloser(buf)
 	mockLayer2.EXPECT().Uncompressed().Return(rc, nil)
 
 	layers := []v1.Layer{
@@ -1240,16 +1224,15 @@ func Test_GetFSFromLayers_with_whiteouts_include_whiteout_disabled(t *testing.T)
 }
 
 func Test_GetFSFromLayers_ignorelist(t *testing.T) {
+	resetMountInfoFile := provideEmptyMountinfoFile()
+	defer resetMountInfoFile()
+
 	ctrl := gomock.NewController(t)
 
-	root, err := ioutil.TempDir("", "layers-test")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.Remove(root)
+	root := t.TempDir()
 	// Write a whiteout path
 	fileContents := []byte("Hello World\n")
-	if err := os.Mkdir(filepath.Join(root, "testdir"), 0775); err != nil {
+	if err := os.Mkdir(filepath.Join(root, "testdir"), 0o775); err != nil {
 		t.Fatal(err)
 	}
 
@@ -1266,7 +1249,7 @@ func Test_GetFSFromLayers_ignorelist(t *testing.T) {
 
 			hdr := &tar.Header{
 				Name: f,
-				Mode: 0644,
+				Mode: 0o644,
 				Size: int64(len(string(fileContents))),
 			}
 
@@ -1308,7 +1291,7 @@ func Test_GetFSFromLayers_ignorelist(t *testing.T) {
 
 	f(layerFiles, tw)
 
-	rc := ioutil.NopCloser(buf)
+	rc := io.NopCloser(buf)
 	mockLayer.EXPECT().Uncompressed().Return(rc, nil)
 
 	layers := []v1.Layer{
@@ -1338,7 +1321,7 @@ func Test_GetFSFromLayers_ignorelist(t *testing.T) {
 	defaultIgnoreList = append(defaultIgnoreList, IgnoreListEntry{
 		Path: filepath.Join(root, "testdir"),
 	})
-	if err := os.Mkdir(filepath.Join(root, "testdir"), 0775); err != nil {
+	if err := os.Mkdir(filepath.Join(root, "testdir"), 0o775); err != nil {
 		t.Fatal(err)
 	}
 
@@ -1362,7 +1345,7 @@ func Test_GetFSFromLayers_ignorelist(t *testing.T) {
 
 	f(layerFiles, tw)
 
-	rc = ioutil.NopCloser(buf)
+	rc = io.NopCloser(buf)
 	mockLayer.EXPECT().Uncompressed().Return(rc, nil)
 
 	layers = []v1.Layer{
@@ -1388,11 +1371,7 @@ func Test_GetFSFromLayers_ignorelist(t *testing.T) {
 func Test_GetFSFromLayers(t *testing.T) {
 	ctrl := gomock.NewController(t)
 
-	root, err := ioutil.TempDir("", "layers-test")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.Remove(root)
+	root := t.TempDir()
 
 	opts := []FSOpt{
 		// I'd rather use the real func (util.ExtractFile)
@@ -1413,7 +1392,7 @@ func Test_GetFSFromLayers(t *testing.T) {
 
 		hdr := &tar.Header{
 			Name: f,
-			Mode: 0644,
+			Mode: 0o644,
 			Size: int64(len("Hello world\n")),
 		}
 
@@ -1433,7 +1412,7 @@ func Test_GetFSFromLayers(t *testing.T) {
 	mockLayer := mockv1.NewMockLayer(ctrl)
 	mockLayer.EXPECT().MediaType().Return(types.OCILayer, nil)
 
-	rc := ioutil.NopCloser(buf)
+	rc := io.NopCloser(buf)
 	mockLayer.EXPECT().Uncompressed().Return(rc, nil)
 
 	layers := []v1.Layer{
@@ -1483,7 +1462,7 @@ func TestInitIgnoreList(t *testing.T) {
 	mountInfo := `36 35 98:0 /kaniko /test/kaniko rw,noatime master:1 - ext3 /dev/root rw,errors=continue
 36 35 98:0 /proc /test/proc rw,noatime master:1 - ext3 /dev/root rw,errors=continue
 `
-	mFile, err := ioutil.TempFile("", "mountinfo")
+	mFile, err := os.CreateTemp("", "mountinfo")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -1491,9 +1470,9 @@ func TestInitIgnoreList(t *testing.T) {
 	if _, err := mFile.WriteString(mountInfo); err != nil {
 		t.Fatal(err)
 	}
-	config.IgnoreListPath = mFile.Name()
+	config.MountInfoPath = mFile.Name()
 	defer func() {
-		config.IgnoreListPath = constants.IgnoreListPath
+		config.MountInfoPath = constants.MountInfoPath
 	}()
 
 	expected := []IgnoreListEntry{
@@ -1522,7 +1501,7 @@ func TestInitIgnoreList(t *testing.T) {
 	original := append([]IgnoreListEntry{}, ignorelist...)
 	defer func() { ignorelist = original }()
 
-	err = InitIgnoreList(true)
+	err = InitIgnoreList()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -1536,16 +1515,11 @@ func TestInitIgnoreList(t *testing.T) {
 }
 
 func Test_setFileTimes(t *testing.T) {
-	testDir, err := ioutil.TempDir("", "")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	defer os.RemoveAll(testDir)
+	testDir := t.TempDir()
 
 	p := filepath.Join(testDir, "foo.txt")
 
-	if err := ioutil.WriteFile(p, []byte("meow"), 0777); err != nil {
+	if err := os.WriteFile(p, []byte("meow"), 0o777); err != nil {
 		t.Fatal(err)
 	}
 
