@@ -18,8 +18,8 @@ package testutil
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
+	"os/user"
 	"path/filepath"
 	"reflect"
 	"testing"
@@ -34,11 +34,38 @@ func SetupFiles(path string, files map[string]string) error {
 		if err := os.MkdirAll(filepath.Dir(path), 0750); err != nil {
 			return err
 		}
-		if err := ioutil.WriteFile(path, []byte(c), 0644); err != nil {
+		if err := os.WriteFile(path, []byte(c), 0644); err != nil {
 			return err
 		}
 	}
 	return nil
+}
+
+type CurrentUser struct {
+	*user.User
+
+	PrimaryGroup string
+}
+
+func GetCurrentUser(t *testing.T) CurrentUser {
+	currentUser, err := user.Current()
+	if err != nil {
+		t.Fatalf("Cannot get current user: %s", err)
+	}
+	groups, err := currentUser.GroupIds()
+	if err != nil || len(groups) == 0 {
+		t.Fatalf("Cannot get groups for current user: %s", err)
+	}
+	primaryGroupObj, err := user.LookupGroupId(groups[0])
+	if err != nil {
+		t.Fatalf("Could not lookup name of group %s: %s", groups[0], err)
+	}
+	primaryGroup := primaryGroupObj.Name
+
+	return CurrentUser{
+		User:         currentUser,
+		PrimaryGroup: primaryGroup,
+	}
 }
 
 func CheckDeepEqual(t *testing.T, expected, actual interface{}) {
@@ -79,7 +106,7 @@ func checkErr(shouldErr bool, err error) error {
 		return fmt.Errorf("Expected error, but returned none")
 	}
 	if err != nil && !shouldErr {
-		return fmt.Errorf("Unexpected error: %s", err)
+		return fmt.Errorf("Unexpected error: %w", err)
 	}
 	return nil
 }
