@@ -17,16 +17,20 @@ limitations under the License.
 package buildcontext
 
 import (
+	"context"
+	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
 
+	kConfig "github.com/GoogleContainerTools/kaniko/pkg/config"
 	"github.com/GoogleContainerTools/kaniko/pkg/constants"
 	"github.com/GoogleContainerTools/kaniko/pkg/util"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/s3"
-	"github.com/aws/aws-sdk-go/service/s3/s3manager"
+	"github.com/GoogleContainerTools/kaniko/pkg/util/bucket"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/config"
+	s3manager "github.com/aws/aws-sdk-go-v2/feature/s3/manager"
+	"github.com/aws/aws-sdk-go-v2/service/s3"
 )
 
 // S3 unifies calls to download and unpack the build context.
@@ -36,27 +40,37 @@ type S3 struct {
 
 // UnpackTarFromBuildContext download and untar a file from s3
 func (s *S3) UnpackTarFromBuildContext() (string, error) {
-	bucket, item := util.GetBucketAndItem(s.context)
-	option := session.Options{
-		SharedConfigState: session.SharedConfigEnable,
+	bucket, item, err := bucket.GetNameAndFilepathFromURI(s.context)
+	if err != nil {
+		return "", fmt.Errorf("getting bucketname and filepath from context: %w", err)
 	}
+
 	endpoint := os.Getenv(constants.S3EndpointEnv)
 	forcePath := false
 	if strings.ToLower(os.Getenv(constants.S3ForcePathStyle)) == "true" {
 		forcePath = true
 	}
-	if endpoint != "" {
-		option.Config = aws.Config{
-			Endpoint:         aws.String(endpoint),
-			S3ForcePathStyle: aws.Bool(forcePath),
+
+	customResolver := aws.EndpointResolverWithOptionsFunc(func(service, region string, options ...interface{}) (aws.Endpoint, error) {
+		if endpoint != "" {
+			return aws.Endpoint{
+				URL: endpoint,
+			}, nil
 		}
-	}
-	sess, err := session.NewSessionWithOptions(option)
+		return aws.Endpoint{}, &aws.EndpointNotFoundError{}
+	})
+
+	cfg, err := config.LoadDefaultConfig(context.TODO(), config.WithEndpointResolverWithOptions(customResolver))
 	if err != nil {
 		return bucket, err
 	}
-	downloader := s3manager.NewDownloader(sess)
-	directory := constants.BuildContextDir
+	client := s3.NewFromConfig(cfg, func(options *s3.Options) {
+		if endpoint != "" {
+			options.UsePathStyle = forcePath
+		}
+	})
+	downloader := s3manager.NewDownloader(client)
+	directory := kConfig.BuildContextDir
 	tarPath := filepath.Join(directory, constants.ContextTar)
 	if err := os.MkdirAll(directory, 0750); err != nil {
 		return directory, err
@@ -65,7 +79,7 @@ func (s *S3) UnpackTarFromBuildContext() (string, error) {
 	if err != nil {
 		return directory, err
 	}
-	_, err = downloader.Download(file,
+	_, err = downloader.Download(context.TODO(), file,
 		&s3.GetObjectInput{
 			Bucket: aws.String(bucket),
 			Key:    aws.String(item),
