@@ -28,10 +28,10 @@ type OnBuildCommand struct {
 	cmd *instructions.OnbuildCommand
 }
 
-//ExecuteCommand adds the specified expression in Onbuild to the config
+// ExecuteCommand adds the specified expression in Onbuild to the config
 func (o *OnBuildCommand) ExecuteCommand(config *v1.Config, buildArgs *dockerfile.BuildArgs) error {
-	logrus.Info("cmd: ONBUILD")
-	logrus.Infof("args: %s", o.cmd.Expression)
+	logrus.Info("Cmd: ONBUILD")
+	logrus.Infof("Args: %s", o.cmd.Expression)
 	if config.OnBuild == nil {
 		config.OnBuild = []string{o.cmd.Expression}
 	} else {
