@@ -28,20 +28,21 @@ import (
 
 type RunMarkerCommand struct {
 	BaseCommand
-	cmd   *instructions.RunCommand
-	Files []string
+	cmd      *instructions.RunCommand
+	Files    []string
+	shdCache bool
 }
 
 func (r *RunMarkerCommand) ExecuteCommand(config *v1.Config, buildArgs *dockerfile.BuildArgs) error {
 	// run command `touch filemarker`
-	logrus.Debugf("using new RunMarker command")
+	logrus.Debugf("Using new RunMarker command")
 	prevFilesMap, _ := util.GetFSInfoMap("/", map[string]os.FileInfo{})
 	if err := runCommandInExec(config, buildArgs, r.cmd); err != nil {
 		return err
 	}
 	_, r.Files = util.GetFSInfoMap("/", prevFilesMap)
 
-	logrus.Debugf("files changed %s", r.Files)
+	logrus.Debugf("Files changed %s", r.Files)
 	return nil
 }
 
@@ -55,6 +56,10 @@ func (r *RunMarkerCommand) FilesToSnapshot() []string {
 }
 
 func (r *RunMarkerCommand) ProvidesFilesToSnapshot() bool {
+	return true
+}
+
+func (r *RunMarkerCommand) IsArgsEnvsRequiredInCache() bool {
 	return true
 }
 
@@ -77,7 +82,7 @@ func (r *RunMarkerCommand) RequiresUnpackedFS() bool {
 }
 
 func (r *RunMarkerCommand) ShouldCacheOutput() bool {
-	return true
+	return r.shdCache
 }
 
 func (r *RunMarkerCommand) ShouldDetectDeletedFiles() bool {
