@@ -19,9 +19,9 @@ package commands
 import (
 	"github.com/GoogleContainerTools/kaniko/pkg/dockerfile"
 	"github.com/GoogleContainerTools/kaniko/pkg/util"
-	"github.com/docker/docker/pkg/signal"
 	v1 "github.com/google/go-containerregistry/pkg/v1"
 	"github.com/moby/buildkit/frontend/dockerfile/instructions"
+	"github.com/moby/sys/signal"
 	"github.com/sirupsen/logrus"
 )
 
@@ -32,7 +32,7 @@ type StopSignalCommand struct {
 
 // ExecuteCommand handles command processing similar to CMD and RUN,
 func (s *StopSignalCommand) ExecuteCommand(config *v1.Config, buildArgs *dockerfile.BuildArgs) error {
-	logrus.Info("cmd: STOPSIGNAL")
+	logrus.Info("Cmd: STOPSIGNAL")
 
 	// resolve possible environment variables
 	replacementEnvs := buildArgs.ReplacementEnvs(config.Env)
