@@ -17,7 +17,7 @@ limitations under the License.
 package remote
 
 import (
-	"runtime"
+	"fmt"
 	"strings"
 
 	"github.com/GoogleContainerTools/kaniko/pkg/config"
@@ -32,7 +32,8 @@ import (
 )
 
 var (
-	manifestCache = make(map[string]v1.Image)
+	manifestCache   = make(map[string]v1.Image)
+	remoteImageFunc = remote.Image
 )
 
 // RetrieveRemoteImage retrieves the manifest for the specified image from the specified registry
@@ -50,34 +51,38 @@ func RetrieveRemoteImage(image string, opts config.RegistryOptions, customPlatfo
 		return nil, err
 	}
 
-	if ref.Context().RegistryStr() == name.DefaultRegistry {
-		ref, err := normalizeReference(ref, image)
-		if err != nil {
-			return nil, err
-		}
+	if newRegURLs, found := opts.RegistryMaps[ref.Context().RegistryStr()]; found {
+		for _, registryMapping := range newRegURLs {
 
-		for _, registryMirror := range opts.RegistryMirrors {
-			var newReg name.Registry
-			if opts.InsecurePull || opts.InsecureRegistries.Contains(registryMirror) {
-				newReg, err = name.NewRegistry(registryMirror, name.WeakValidation, name.Insecure)
-			} else {
-				newReg, err = name.NewRegistry(registryMirror, name.StrictValidation)
-			}
+			regToMapTo, repositoryPrefix := parseRegistryMapping(registryMapping)
+
+			insecurePull := opts.InsecurePull || opts.InsecureRegistries.Contains(regToMapTo)
+
+			remappedRepository, err := remapRepository(ref.Context(), regToMapTo, repositoryPrefix, insecurePull)
 			if err != nil {
 				return nil, err
 			}
-			ref := setNewRegistry(ref, newReg)
 
-			logrus.Infof("Retrieving image %s from registry mirror %s", ref, registryMirror)
-			remoteImage, err := remote.Image(ref, remoteOptions(registryMirror, opts, customPlatform)...)
-			if err != nil {
-				logrus.Warnf("Failed to retrieve image %s from registry mirror %s: %s. Will try with the next mirror, or fallback to the default registry.", ref, registryMirror, err)
+			remappedRef := setNewRepository(ref, remappedRepository)
+
+			logrus.Infof("Retrieving image %s from mapped registry %s", remappedRef, regToMapTo)
+			retryFunc := func() (v1.Image, error) {
+				return remoteImageFunc(remappedRef, remoteOptions(regToMapTo, opts, customPlatform)...)
+			}
+
+			var remoteImage v1.Image
+			if remoteImage, err = util.RetryWithResult(retryFunc, opts.ImageDownloadRetry, 1000); err != nil {
+				logrus.Warnf("Failed to retrieve image %s from remapped registry %s: %s. Will try with the next registry, or fallback to the original registry.", remappedRef, regToMapTo, err)
 				continue
 			}
 
 			manifestCache[image] = remoteImage
 
 			return remoteImage, nil
+		}
+
+		if len(newRegURLs) > 0 && opts.SkipDefaultRegistryFallback {
+			return nil, fmt.Errorf("image not found on any configured mapped registries for %s", ref)
 		}
 	}
 
@@ -92,25 +97,38 @@ func RetrieveRemoteImage(image string, opts config.RegistryOptions, customPlatfo
 
 	logrus.Infof("Retrieving image %s from registry %s", ref, registryName)
 
-	remoteImage, err := remote.Image(ref, remoteOptions(registryName, opts, customPlatform)...)
+	retryFunc := func() (v1.Image, error) {
+		return remoteImageFunc(ref, remoteOptions(registryName, opts, customPlatform)...)
+	}
 
-	if remoteImage != nil {
+	var remoteImage v1.Image
+	if remoteImage, err = util.RetryWithResult(retryFunc, opts.ImageDownloadRetry, 1000); remoteImage != nil {
 		manifestCache[image] = remoteImage
 	}
 
 	return remoteImage, err
 }
 
-// normalizeReference adds the library/ prefix to images without it.
-//
-// It is mostly useful when using a registry mirror that is not able to perform
-// this fix automatically.
-func normalizeReference(ref name.Reference, image string) (name.Reference, error) {
-	if !strings.ContainsRune(image, '/') {
-		return name.ParseReference("library/"+image, name.WeakValidation)
+// remapRepository adds the {repositoryPrefix}/ to the original repo, and normalizes with an additional library/ if necessary
+func remapRepository(repo name.Repository, regToMapTo string, repositoryPrefix string, insecurePull bool) (name.Repository, error) {
+	if insecurePull {
+		return name.NewRepository(repositoryPrefix+repo.RepositoryStr(), name.WithDefaultRegistry(regToMapTo), name.WeakValidation, name.Insecure)
+	} else {
+		return name.NewRepository(repositoryPrefix+repo.RepositoryStr(), name.WithDefaultRegistry(regToMapTo), name.WeakValidation)
 	}
+}
 
-	return ref, nil
+func setNewRepository(ref name.Reference, newRepo name.Repository) name.Reference {
+	switch r := ref.(type) {
+	case name.Tag:
+		r.Repository = newRepo
+		return r
+	case name.Digest:
+		r.Repository = newRepo
+		return r
+	default:
+		return ref
+	}
 }
 
 func setNewRegistry(ref name.Reference, newReg name.Registry) name.Reference {
@@ -127,30 +145,33 @@ func setNewRegistry(ref name.Reference, newReg name.Registry) name.Reference {
 }
 
 func remoteOptions(registryName string, opts config.RegistryOptions, customPlatform string) []remote.Option {
-	tr := util.MakeTransport(opts, registryName)
+	tr, err := util.MakeTransport(opts, registryName)
 
-	// on which v1.Platform is this currently running?
-	platform := currentPlatform(customPlatform)
+	// The MakeTransport function will only return errors if there was a problem
+	// with registry certificates (Verification or mTLS)
+	if err != nil {
+		logrus.Fatalf("Unable to setup transport for registry %q: %v", customPlatform, err)
+	}
 
-	return []remote.Option{remote.WithTransport(tr), remote.WithAuthFromKeychain(creds.GetKeychain()), remote.WithPlatform(platform)}
+	// The platform value has previously been validated.
+	platform, err := v1.ParsePlatform(customPlatform)
+	if err != nil {
+		logrus.Fatalf("Invalid platform %q: %v", customPlatform, err)
+	}
+
+	return []remote.Option{remote.WithTransport(tr), remote.WithAuthFromKeychain(creds.GetKeychain()), remote.WithPlatform(*platform)}
 }
 
-// CurrentPlatform returns the v1.Platform on which the code runs
-func currentPlatform(customPlatform string) v1.Platform {
-	if customPlatform != "" {
-		customPlatformArray := strings.Split(customPlatform, "/")
-		imagePlatform := v1.Platform{}
-		imagePlatform.OS = customPlatformArray[0]
-		if len(customPlatformArray) > 1 {
-			imagePlatform.Architecture = customPlatformArray[1]
-			if len(customPlatformArray) > 2 {
-				imagePlatform.Variant = customPlatformArray[2]
-			}
-		}
-		return imagePlatform
+// Parse the registry mapping
+// example: regMapping = "registry.example.com/subdir1/subdir2" will return registry.example.com and subdir1/subdir2/
+func parseRegistryMapping(regMapping string) (string, string) {
+	// Split the registry mapping by first slash
+	regURL, repositoryPrefix, _ := strings.Cut(regMapping, "/")
+
+	// Normalize with a trailing slash if not empty
+	if repositoryPrefix != "" && !strings.HasSuffix(repositoryPrefix, "/") {
+		repositoryPrefix = repositoryPrefix + "/"
 	}
-	return v1.Platform{
-		OS:           runtime.GOOS,
-		Architecture: runtime.GOARCH,
-	}
+
+	return regURL, repositoryPrefix
 }
