@@ -17,16 +17,21 @@ limitations under the License.
 package remote
 
 import (
+	"errors"
 	"testing"
 
 	"github.com/GoogleContainerTools/kaniko/pkg/config"
 	"github.com/google/go-containerregistry/pkg/name"
 	v1 "github.com/google/go-containerregistry/pkg/v1"
+	"github.com/google/go-containerregistry/pkg/v1/remote"
 	"github.com/google/go-containerregistry/pkg/v1/types"
 )
 
+const image string = "debian"
+
 // mockImage mocks the v1.Image interface
-type mockImage struct{}
+type mockImage struct {
+}
 
 func (m *mockImage) ConfigFile() (*v1.ConfigFile, error) {
 	return nil, nil
@@ -76,22 +81,95 @@ func (m *mockImage) Size() (int64, error) {
 	return 0, nil
 }
 
-func Test_normalizeReference(t *testing.T) {
-	image := "debian"
-	expected := "index.docker.io/library/debian:latest"
-
-	ref, err := name.ParseReference(image)
-	if err != nil {
-		t.Fatal(err)
+func Test_remapRepository(t *testing.T) {
+	tests := []struct {
+		name                string
+		repository          string
+		newRegistry         string
+		newRepositoryPrefix string
+		expectedRepository  string
+	}{
+		{
+			name:                "Test case 1",
+			repository:          "debian",
+			newRegistry:         "newreg.io",
+			newRepositoryPrefix: "",
+			expectedRepository:  "newreg.io/library/debian",
+		},
+		{
+			name:                "Test case 2",
+			repository:          "docker.io/debian",
+			newRegistry:         "newreg.io",
+			newRepositoryPrefix: "",
+			expectedRepository:  "newreg.io/library/debian",
+		},
+		{
+			name:                "Test case 3",
+			repository:          "index.docker.io/debian",
+			newRegistry:         "newreg.io",
+			newRepositoryPrefix: "",
+			expectedRepository:  "newreg.io/library/debian",
+		},
+		{
+			name:                "Test case 4",
+			repository:          "oldreg.io/debian",
+			newRegistry:         "newreg.io",
+			newRepositoryPrefix: "",
+			expectedRepository:  "newreg.io/debian",
+		},
+		{
+			name:                "Test case 5",
+			repository:          "debian",
+			newRegistry:         "newreg.io",
+			newRepositoryPrefix: "subdir1/subdir2/",
+			expectedRepository:  "newreg.io/subdir1/subdir2/library/debian",
+		},
+		{
+			name:                "Test case 6",
+			repository:          "library/debian",
+			newRegistry:         "newreg.io",
+			newRepositoryPrefix: "",
+			expectedRepository:  "newreg.io/library/debian",
+		},
+		{
+			name:                "Test case 7",
+			repository:          "library/debian",
+			newRegistry:         "newreg.io",
+			newRepositoryPrefix: "subdir1/subdir2/",
+			expectedRepository:  "newreg.io/subdir1/subdir2/library/debian",
+		},
+		{
+			name:                "Test case 8",
+			repository:          "namespace/debian",
+			newRegistry:         "newreg.io",
+			newRepositoryPrefix: "",
+			expectedRepository:  "newreg.io/namespace/debian",
+		},
+		{
+			name:                "Test case 9",
+			repository:          "namespace/debian",
+			newRegistry:         "newreg.io",
+			newRepositoryPrefix: "subdir1/subdir2/",
+			expectedRepository:  "newreg.io/subdir1/subdir2/namespace/debian",
+		},
+		// Add more test cases here
 	}
 
-	ref2, err := normalizeReference(ref, image)
-	if err != nil {
-		t.Fatal(err)
-	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			repo, err := name.NewRepository(tt.repository)
+			if err != nil {
+				t.Fatal(err)
+			}
+			repo2, err := remapRepository(repo, tt.newRegistry, tt.newRepositoryPrefix, false)
+			if err != nil {
+				t.Fatal(err)
+			}
 
-	if ref2.Name() != ref.Name() || ref2.Name() != expected {
-		t.Errorf("%s should have been normalized to %s, got %s", ref2.Name(), expected, ref.Name())
+			if repo2.Name() != tt.expectedRepository {
+				t.Errorf("%s should have been normalized to %s, got %s", repo.Name(), tt.expectedRepository, repo2.Name())
+			}
+		})
 	}
 }
 
@@ -106,5 +184,129 @@ func Test_RetrieveRemoteImage_manifestCache(t *testing.T) {
 
 	if image, err := RetrieveRemoteImage(nonExistingImageName, config.RegistryOptions{}, ""); image == nil || err != nil {
 		t.Fatal("Expected call to succeed because there is a manifest for this image in the cache.")
+	}
+}
+
+func Test_RetrieveRemoteImage_skipFallback(t *testing.T) {
+	registryMirror := "some-registry"
+
+	opts := config.RegistryOptions{
+		RegistryMaps:                map[string][]string{name.DefaultRegistry: {registryMirror}},
+		SkipDefaultRegistryFallback: false,
+	}
+
+	remoteImageFunc = func(ref name.Reference, options ...remote.Option) (v1.Image, error) {
+		if ref.Context().Registry.Name() == registryMirror {
+			return nil, errors.New("no image found")
+		}
+
+		return &mockImage{}, nil
+	}
+
+	if _, err := RetrieveRemoteImage(image, opts, ""); err != nil {
+		t.Fatalf("Expected call to succeed because fallback to default registry")
+	}
+
+	opts.SkipDefaultRegistryFallback = true
+	//clean cached image
+	manifestCache = make(map[string]v1.Image)
+
+	if _, err := RetrieveRemoteImage(image, opts, ""); err == nil {
+		t.Fatal("Expected call to fail because fallback to default registry is skipped")
+	}
+}
+
+func Test_RetryRetrieveRemoteImageSucceeds(t *testing.T) {
+	opts := config.RegistryOptions{
+		ImageDownloadRetry: 2,
+	}
+	attempts := 0
+	remoteImageFunc = func(ref name.Reference, options ...remote.Option) (v1.Image, error) {
+		if attempts < 2 {
+			attempts++
+			return nil, errors.New("no image found")
+		}
+		return &mockImage{}, nil
+	}
+
+	// Clean cached image
+	manifestCache = make(map[string]v1.Image)
+
+	if _, err := RetrieveRemoteImage(image, opts, ""); err != nil {
+		t.Fatal("Expected call to succeed because of retry")
+	}
+}
+
+func Test_NoRetryRetrieveRemoteImageFails(t *testing.T) {
+	opts := config.RegistryOptions{
+		ImageDownloadRetry: 0,
+	}
+	attempts := 0
+	remoteImageFunc = func(ref name.Reference, options ...remote.Option) (v1.Image, error) {
+		if attempts < 1 {
+			attempts++
+			return nil, errors.New("no image found")
+		}
+		return &mockImage{}, nil
+	}
+
+	// Clean cached image
+	manifestCache = make(map[string]v1.Image)
+
+	if _, err := RetrieveRemoteImage(image, opts, ""); err == nil {
+		t.Fatal("Expected call to fail because there is no retry")
+	}
+}
+
+func Test_ParseRegistryMapping(t *testing.T) {
+	tests := []struct {
+		name                     string
+		registryMapping          string
+		expectedRegistry         string
+		expectedRepositoryPrefix string
+	}{
+		{
+			name:                     "Test case 1",
+			registryMapping:          "registry.example.com/subdir",
+			expectedRegistry:         "registry.example.com",
+			expectedRepositoryPrefix: "subdir/",
+		},
+		{
+			name:                     "Test case 2",
+			registryMapping:          "registry.example.com/subdir/",
+			expectedRegistry:         "registry.example.com",
+			expectedRepositoryPrefix: "subdir/",
+		},
+		{
+			name:                     "Test case 3",
+			registryMapping:          "registry.example.com/subdir1/subdir2",
+			expectedRegistry:         "registry.example.com",
+			expectedRepositoryPrefix: "subdir1/subdir2/",
+		},
+		{
+			name:                     "Test case 4",
+			registryMapping:          "registry.example.com",
+			expectedRegistry:         "registry.example.com",
+			expectedRepositoryPrefix: "",
+		},
+		{
+			name:                     "Test case 5",
+			registryMapping:          "registry.example.com/",
+			expectedRegistry:         "registry.example.com",
+			expectedRepositoryPrefix: "",
+		},
+		// Add more test cases here
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			registry, repositoryPrefix := parseRegistryMapping(tt.registryMapping)
+			if registry != tt.expectedRegistry {
+				t.Errorf("Expected registry: %s, but got: %s", tt.expectedRegistry, registry)
+			}
+			if repositoryPrefix != tt.expectedRepositoryPrefix {
+				t.Errorf("Expected repoPrefix: %s, but got: %s", tt.expectedRepositoryPrefix, repositoryPrefix)
+			}
+		})
 	}
 }
