@@ -13,6 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package snapshot
 
 import (
@@ -61,8 +62,8 @@ func Test_CacheKey(t *testing.T) {
 	}
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			lm1 := LayeredMap{layers: []map[string]string{test.map1}}
-			lm2 := LayeredMap{layers: []map[string]string{test.map2}}
+			lm1 := LayeredMap{adds: []map[string]string{test.map1}, deletes: []map[string]struct{}{nil, nil}}
+			lm2 := LayeredMap{adds: []map[string]string{test.map2}, deletes: []map[string]struct{}{nil, nil}}
 			k1, err := lm1.Key()
 			if err != nil {
 				t.Fatalf("error getting key for map 1: %v", err)
@@ -76,4 +77,68 @@ func Test_CacheKey(t *testing.T) {
 			}
 		})
 	}
+}
+
+func Test_FlattenPaths(t *testing.T) {
+	layers := []map[string]string{
+		{
+			"a": "2",
+			"b": "3",
+		},
+		{
+			"b": "5",
+			"c": "6",
+		},
+		{
+			"a": "8",
+		},
+	}
+
+	whiteouts := []map[string]struct{}{
+		{
+			"a": {}, // delete a
+		},
+		{
+			"b": {}, // delete b
+		},
+		{
+			"c": {}, // delete c
+		},
+	}
+
+	lm := LayeredMap{
+		adds:    []map[string]string{layers[0]},
+		deletes: []map[string]struct{}{whiteouts[0]}}
+
+	paths := lm.GetCurrentPaths()
+
+	assertPath := func(f string, exists bool) {
+		_, ok := paths[f]
+		if exists && !ok {
+			t.Fatalf("expected path '%s' to be present.", f)
+		} else if !exists && ok {
+			t.Fatalf("expected path '%s' not to be present.", f)
+		}
+	}
+
+	assertPath("a", false)
+	assertPath("b", true)
+
+	lm = LayeredMap{
+		adds:    []map[string]string{layers[0], layers[1]},
+		deletes: []map[string]struct{}{whiteouts[0], whiteouts[1]}}
+	paths = lm.GetCurrentPaths()
+
+	assertPath("a", false)
+	assertPath("b", false)
+	assertPath("c", true)
+
+	lm = LayeredMap{
+		adds:    []map[string]string{layers[0], layers[1], layers[2]},
+		deletes: []map[string]struct{}{whiteouts[0], whiteouts[1], whiteouts[2]}}
+	paths = lm.GetCurrentPaths()
+
+	assertPath("a", true)
+	assertPath("b", false)
+	assertPath("c", false)
 }
